% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions-documentation.R
\name{additionalDoc}
\alias{additionalDoc}
\title{Common documents}
\arguments{
\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}

\item{...}{The additional request parameters that may not present
in the function parameters.}
}
\description{
Common documents
}
