% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions-documentation.R
\name{CommonDoc}
\alias{CommonDoc}
\title{Common documents}
\arguments{
\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object (see below)[optional]}

\item{NextToken}{Characters, The token for the next page of results[optional]}

\item{nextToken}{Characters, The token for the next page of results[optional]}
}
\description{
Common documents
}
