% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_accept_vpc_peering_connection}
\alias{ec2_accept_vpc_peering_connection}
\title{Accept Vpc Peering Connection}
\usage{
ec2_accept_vpc_peering_connection(
  DryRun = NULL,
  VpcPeeringConnectionId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{VpcPeeringConnectionId}{String. The ID of the VPC peering connection. You must specify this parameter in the request. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Accept a VPC peering connection request. To accept a request, the VPC
peering connection must be in the \code{pending-acceptance} state, and you
must be the owner of the peer VPC. Use DescribeVpcPeeringConnections to
view your outstanding VPC peering connection requests.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{VpcPeeringConnectionId}{

The ID of the VPC peering connection. You must specify this parameter in
the request.
}

