% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_address}
\alias{ec2_associate_address}
\title{Associate Address}
\usage{
ec2_associate_address(
  AllocationId = NULL,
  InstanceId = NULL,
  PublicIp = NULL,
  AllowReassociation = NULL,
  DryRun = NULL,
  NetworkInterfaceId = NULL,
  PrivateIpAddress = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AllocationId}{String. [EC2-VPC] The allocation ID. This is required for EC2-VPC. (optional)}

\item{InstanceId}{String. The ID of the instance.(optional)}

\item{PublicIp}{String. [EC2-Classic] The Elastic IP address to associate with the instance.(optional)}

\item{AllowReassociation}{Boolean. [EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address...(see below)(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{NetworkInterfaceId}{String. [EC2-VPC] The ID of the network interface.(optional)}

\item{PrivateIpAddress}{String. [EC2-VPC] The primary or secondary private IP address to associate with the Elastic IP address.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Associates an Elastic IP address, or carrier IP address (for instances
that are in subnets in Wavelength Zones) with an instance or a network
interface. Before you can use an Elastic IP address, you must allocate
it to your account.
}
\section{AllocationId}{

[EC2-VPC] The allocation ID. This is required for EC2-VPC.
}

\section{InstanceId}{

The ID of the instance. The instance must have exactly one attached
network interface. For EC2-VPC, you can specify either the instance ID
or the network interface ID, but not both. For EC2-Classic, you must
specify an instance ID and the instance must be in the running state.
}

\section{PublicIp}{

[EC2-Classic] The Elastic IP address to associate with the instance.
This is required for EC2-Classic.
}

\section{AllowReassociation}{

[EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow
an Elastic IP address that is already associated with an instance or
network interface to be reassociated with the specified instance or
network interface. Otherwise, the operation fails. In a VPC in an
EC2-VPC-only account, reassociation is automatic, therefore you can
specify false to ensure the operation fails if the Elastic IP address is
already associated with another resource.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{NetworkInterfaceId}{

[EC2-VPC] The ID of the network interface. If the instance has more
than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network
interface ID, but not both.
}

\section{PrivateIpAddress}{

[EC2-VPC] The primary or secondary private IP address to associate
with the Elastic IP address. If no private IP address is specified, the
Elastic IP address is associated with the primary private IP address.
}

