% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_iam_instance_profile}
\alias{ec2_associate_iam_instance_profile}
\title{Associate Iam Instance Profile}
\usage{
ec2_associate_iam_instance_profile(
  IamInstanceProfile,
  InstanceId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{IamInstanceProfile}{Object. The IAM instance profile.}

\item{InstanceId}{String. The ID of the instance.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Associates an IAM instance profile with a running or stopped instance.
You cannot associate more than one IAM instance profile with an
instance.
}
\section{IamInstanceProfile}{

The IAM instance profile.
}

\section{InstanceId}{

The ID of the instance.
}

