% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_vpc_cidr_block}
\alias{ec2_associate_vpc_cidr_block}
\title{Associate Vpc Cidr Block}
\usage{
ec2_associate_vpc_cidr_block(
  VpcId,
  AmazonProvidedIpv6CidrBlock = NULL,
  CidrBlock = NULL,
  Ipv6CidrBlockNetworkBorderGroup = NULL,
  Ipv6Pool = NULL,
  Ipv6CidrBlock = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcId}{String. The ID of the VPC.}

\item{AmazonProvidedIpv6CidrBlock}{Boolean. Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC.(optional)}

\item{CidrBlock}{String. An IPv4 CIDR block to associate with the VPC. (optional)}

\item{Ipv6CidrBlockNetworkBorderGroup}{String. The name of the location from which we advertise the IPV6 CIDR block.(optional)}

\item{Ipv6Pool}{String. The ID of an IPv6 address pool from which to allocate the IPv6 CIDR block. (optional)}

\item{Ipv6CidrBlock}{String. An IPv6 CIDR block from the IPv6 address pool.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Associates a CIDR block with your VPC. You can associate a secondary
IPv4 CIDR block, an Amazon-provided IPv6 CIDR block, or an IPv6 CIDR
block from an IPv6 address pool that you provisioned through bring your
own IP addresses
(\href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-byoip.html}{BYOIP}).
The IPv6 CIDR block size is fixed at /56.
}
\section{VpcId}{

The ID of the VPC.
}

\section{AmazonProvidedIpv6CidrBlock}{

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for
the VPC. You cannot specify the range of IPv6 addresses, or the size of
the CIDR block.
}

\section{CidrBlock}{

An IPv4 CIDR block to associate with the VPC.
}

\section{Ipv6CidrBlockNetworkBorderGroup}{

The name of the location from which we advertise the IPV6 CIDR block.
Use this parameter to limit the CIDR block to this location.

You must set \code{AmazonProvidedIpv6CidrBlock} to \code{true} to use this
parameter.

You can have one IPv6 CIDR block association per network border group.
}

\section{Ipv6Pool}{

The ID of an IPv6 address pool from which to allocate the IPv6 CIDR
block.
}

\section{Ipv6CidrBlock}{

An IPv6 CIDR block from the IPv6 address pool. You must also specify
\code{Ipv6Pool} in the request.

To let Amazon choose the IPv6 CIDR block for you, omit this parameter.
}

