% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_capacity_reservation}
\alias{ec2_cancel_capacity_reservation}
\title{Cancel Capacity Reservation}
\usage{
ec2_cancel_capacity_reservation(
  CapacityReservationId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{CapacityReservationId}{String. The ID of the Capacity Reservation to be cancelled.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels the specified Capacity Reservation, releases the reserved
capacity, and changes the Capacity Reservation\'s state to \code{cancelled}.
}
\section{CapacityReservationId}{

The ID of the Capacity Reservation to be cancelled.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

