% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_conversion_task}
\alias{ec2_cancel_conversion_task}
\title{Cancel Conversion Task}
\usage{
ec2_cancel_conversion_task(
  ConversionTaskId,
  DryRun = NULL,
  ReasonMessage = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ConversionTaskId}{String. The ID of the conversion task.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{ReasonMessage}{String. The reason for canceling the conversion task. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels an active conversion task. The task can be the import of an
instance or volume. The action removes all artifacts of the conversion,
including a partially uploaded volume or instance. If the conversion is
complete or is in the process of transferring the final disk image, the
command fails and returns an exception.
}
\section{ConversionTaskId}{

The ID of the conversion task.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{ReasonMessage}{

The reason for canceling the conversion task.
}

