% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_export_task}
\alias{ec2_cancel_export_task}
\title{Cancel Export Task}
\usage{
ec2_cancel_export_task(ExportTaskId, simplify = TRUE, others = list())
}
\arguments{
\item{ExportTaskId}{String. The ID of the export task. This is the ID returned by \code{CreateInstanceExportTask}.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels an active export task. The request removes all artifacts of the
export, including any partially-created Amazon S3 objects. If the export
task is complete or is in the process of transferring the final disk
image, the command fails and returns an error.
}
\section{ExportTaskId}{

The ID of the export task. This is the ID returned by
\code{CreateInstanceExportTask}.
}

