% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_reserved_instances_listing}
\alias{ec2_cancel_reserved_instances_listing}
\title{Cancel Reserved Instances Listing}
\usage{
ec2_cancel_reserved_instances_listing(
  ReservedInstancesListingId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ReservedInstancesListingId}{Character. The ID of the Reserved Instance listing.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Cancel Reserved Instances Listing
}
\section{ReservedInstancesListingId}{

The ID of
the Reserved Instance listing.
}

