% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_spot_fleet_requests}
\alias{ec2_cancel_spot_fleet_requests}
\title{Cancel Spot Fleet Requests}
\usage{
ec2_cancel_spot_fleet_requests(
  SpotFleetRequestId,
  TerminateInstances,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{SpotFleetRequestId}{List or named list. The IDs of the Spot Fleet requests.}

\item{TerminateInstances}{Boolean. Indicates whether to terminate instances for a Spot Fleet request if it is canceled successfully.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels the specified Spot Fleet requests.
}
\section{SpotFleetRequestId}{

The IDs of the Spot Fleet requests.
}

\section{TerminateInstances}{

Indicates whether to terminate instances for a Spot Fleet request if it
is canceled successfully.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

