% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_spot_instance_requests}
\alias{ec2_cancel_spot_instance_requests}
\title{Cancel Spot Instance Requests}
\usage{
ec2_cancel_spot_instance_requests(
  SpotInstanceRequestId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{SpotInstanceRequestId}{List or named list. One or more Spot Instance request IDs.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels one or more Spot Instance requests.
}
\section{SpotInstanceRequestId}{

One or more Spot Instance request IDs.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

