% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_default_vpc}
\alias{ec2_create_default_vpc}
\title{Create Default Vpc}
\usage{
ec2_create_default_vpc(DryRun = NULL, simplify = TRUE, others = list())
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a default VPC with a size \verb{/16} IPv4 CIDR block and a default
subnet in each Availability Zone. For more information about the
components of a default VPC, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/default-vpc.html}{Default VPC and Default Subnets}
in the \emph{Amazon Virtual Private Cloud User Guide}. You cannot specify the
components of the default VPC yourself.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

