% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_key_pair}
\alias{ec2_create_key_pair}
\title{Create Key Pair}
\usage{
ec2_create_key_pair(
  KeyName,
  DryRun = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{KeyName}{String. A unique name for the key pair.  Constraints: Up to 255 ASCII characters}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{TagSpecification}{List or named list. The tags to apply to the new key pair. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a 2048-bit RSA key pair with the specified name. Amazon EC2
stores the public key and displays the private key for you to save to a
file. The private key is returned as an unencrypted PEM encoded PKCS\#1
private key. If a key with the specified name already exists, Amazon EC2
returns an error.
}
\section{KeyName}{

A unique name for the key pair.

Constraints: Up to 255 ASCII characters
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{TagSpecification}{

The tags to apply to the new key pair.
}

