% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_managed_prefix_list}
\alias{ec2_create_managed_prefix_list}
\title{Create Managed Prefix List}
\usage{
ec2_create_managed_prefix_list(
  PrefixListName,
  MaxEntries,
  AddressFamily,
  DryRun = NULL,
  Entry = NULL,
  TagSpecification = NULL,
  ClientToken = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{PrefixListName}{Character.   A name for the prefix list.}

\item{MaxEntries}{Integer. The maximum number of entries for the prefix list.}

\item{AddressFamily}{Character.   The IP address type.  Valid Values: \code{IPv4} \| \code{IPv6}}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{Entry}{List. One or more entries for the prefix list.[optional]}

\item{TagSpecification}{List. The tags to apply to the prefix list during creation.[optional]}

\item{ClientToken}{Character.   Unique, case-sensitive identifier you provide to ensure the idempotency of the request.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Create Managed Prefix List
}
\section{PrefixListName}{


A name for the prefix list.

Constraints: Up to 255 characters in length. The name cannot start with
\code{com.amazonaws}.
}

\section{MaxEntries}{

The maximum number of entries for the prefix
list.
}

\section{AddressFamily}{


The IP address type.

Valid Values: \code{IPv4} \| \code{IPv6}
}

\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Entry}{

One or more entries for the
prefix list.
}

\section{TagSpecification}{

The tags to apply to the prefix list
during creation.
}

\section{ClientToken}{


Unique, case-sensitive identifier you provide to ensure the idempotency
of the request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.

Constraints: Up to 255 UTF-8 characters in length.
}

