% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_placement_group}
\alias{ec2_create_placement_group}
\title{Create Placement Group}
\usage{
ec2_create_placement_group(
  DryRun = NULL,
  GroupName = NULL,
  Strategy = NULL,
  PartitionCount = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{GroupName}{String. A name for the placement group.(optional)}

\item{Strategy}{String. The placement strategy. (optional)}

\item{PartitionCount}{Integer. The number of partitions. Valid only when \strong{Strategy} is set to \code{partition}. (optional)}

\item{TagSpecification}{List or named list. The tags to apply to the new placement group. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a placement group in which to launch instances. The strategy of
the placement group determines how the instances are organized within
the group.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{GroupName}{

A name for the placement group. Must be unique within the scope of your
account for the Region.

Constraints: Up to 255 ASCII characters
}

\section{Strategy}{

The placement strategy.
}

\section{PartitionCount}{

The number of partitions. Valid only when \strong{Strategy} is set to
\code{partition}.
}

\section{TagSpecification}{

The tags to apply to the new placement group.
}

