% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_route_table}
\alias{ec2_create_route_table}
\title{Create Route Table}
\usage{
ec2_create_route_table(
  VpcId,
  DryRun = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcId}{String. The ID of the VPC.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{TagSpecification}{List or named list. The tags to assign to the route table. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a route table for the specified VPC. After you create a route
table, you can add routes and associate the table with a subnet.
}
\section{VpcId}{

The ID of the VPC.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{TagSpecification}{

The tags to assign to the route table.
}

