% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_tags}
\alias{ec2_create_tags}
\title{Create Tags}
\usage{
ec2_create_tags(
  ResourceId,
  Tag,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ResourceId}{List or named list. The IDs of the resources, separated by spaces.}

\item{Tag}{List or named list. The tags.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Adds or overwrites only the specified tags for the specified Amazon EC2
resource or resources. When you specify an existing tag key, the value
is overwritten with the new value. Each resource can have a maximum of
50 tags. Each tag consists of a key and optional value. Tag keys must be
unique per resource.
}
\section{ResourceId}{

The IDs of the resources, separated by spaces.

Constraints: Up to 1000 resource IDs. We recommend breaking up this
request into smaller batches.
}

\section{Tag}{

The tags. The \code{value} parameter is required, but if you don\'t want the
tag to have a value, specify the parameter with no value, and we set the
value to an empty string.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

