% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway}
\alias{ec2_create_transit_gateway}
\title{Create Transit Gateway}
\usage{
ec2_create_transit_gateway(
  Description = NULL,
  Options = NULL,
  TagSpecification = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Description}{Character. A description of the transit gateway.[optional]}

\item{Options}{Object. The transit gateway options.[optional]}

\item{TagSpecification}{List. The tags to apply to the transit gateway.[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Create Transit Gateway
}
\section{Description}{

A
description of the transit gateway.
}

\section{Options}{

The transit
gateway options.
}

\section{TagSpecification}{

The tags to apply to the transit
gateway.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

