% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway_connect}
\alias{ec2_create_transit_gateway_connect}
\title{Create Transit Gateway Connect}
\usage{
ec2_create_transit_gateway_connect(
  TransportTransitGatewayAttachmentId,
  Options,
  TagSpecification = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransportTransitGatewayAttachmentId}{String. The ID of the transit gateway attachment.}

\item{Options}{Object. The Connect attachment options.}

\item{TagSpecification}{List or named list. The tags to apply to the Connect attachment. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a Connect attachment from a specified transit gateway
attachment. A Connect attachment is a GRE-based tunnel attachment that
you can use to establish a connection between a transit gateway and an
appliance.
}
\section{TransportTransitGatewayAttachmentId}{

The ID of the transit gateway attachment. You can specify a VPC
attachment or a AWS Direct Connect attachment.
}

\section{Options}{

The Connect attachment options.
}

\section{TagSpecification}{

The tags to apply to the Connect attachment.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

