% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway_multicast_domain}
\alias{ec2_create_transit_gateway_multicast_domain}
\title{Create Transit Gateway Multicast Domain}
\usage{
ec2_create_transit_gateway_multicast_domain(
  TransitGatewayId,
  Options = NULL,
  TagSpecification = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayId}{Character. The ID of the transit gateway.}

\item{Options}{Object. The options for the transit gateway multicast domain.[optional]}

\item{TagSpecification}{List. The tags for the transit gateway multicast domain.[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Create Transit Gateway Multicast Domain
}
\section{TransitGatewayId}{

The ID of the transit
gateway.
}

\section{Options}{

The options for the transit gateway
multicast domain.
}

\section{TagSpecification}{

The tags for the transit gateway
multicast domain.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

