% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway_route_table}
\alias{ec2_create_transit_gateway_route_table}
\title{Create Transit Gateway Route Table}
\usage{
ec2_create_transit_gateway_route_table(
  TransitGatewayId,
  TagSpecifications = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayId}{String. The ID of the transit gateway.}

\item{TagSpecifications}{List or named list. The tags to apply to the transit gateway route table. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a route table for the specified transit gateway.
}
\section{TransitGatewayId}{

The ID of the transit gateway.
}

\section{TagSpecifications}{

The tags to apply to the transit gateway route table.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

