% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_vpc_endpoint}
\alias{ec2_create_vpc_endpoint}
\title{Create Vpc Endpoint}
\usage{
ec2_create_vpc_endpoint(
  VpcId,
  ServiceName,
  DryRun = NULL,
  VpcEndpointType = NULL,
  PolicyDocument = NULL,
  RouteTableId = NULL,
  SubnetId = NULL,
  SecurityGroupId = NULL,
  ClientToken = NULL,
  PrivateDnsEnabled = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcId}{String. The ID of the VPC in which the endpoint will be used.}

\item{ServiceName}{String. The service name.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{VpcEndpointType}{String. The type of endpoint.  Default: Gateway (optional)}

\item{PolicyDocument}{String. (Interface and gateway endpoints) A policy to attach to the endpoint that controls access...(see below)(optional)}

\item{RouteTableId}{List or named list. (Gateway endpoint) One or more route table IDs. (optional)}

\item{SubnetId}{List or named list. (Interface and Gateway Load Balancer endpoints) The ID of one or more subnets in which to...(see below)(optional)}

\item{SecurityGroupId}{List or named list. (Interface endpoint) The ID of one or more security groups to associate with the endpoint...(see below)(optional)}

\item{ClientToken}{String. Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.(optional)}

\item{PrivateDnsEnabled}{Boolean. (Interface endpoint) Indicates whether to associate a private hosted zone with the specified VPC.(optional)}

\item{TagSpecification}{List or named list. The tags to associate with the endpoint. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a VPC endpoint for a specified service. An endpoint enables you
to create a private connection between your VPC and the service. The
service may be provided by AWS, an AWS Marketplace Partner, or another
AWS account. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/vpc-endpoints.html}{VPC Endpoints}
in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\section{VpcId}{

The ID of the VPC in which the endpoint will be used.
}

\section{ServiceName}{

The service name. To get a list of available services, use the
DescribeVpcEndpointServices request, or get the name from the service
provider.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{VpcEndpointType}{

The type of endpoint.

Default: Gateway
}

\section{PolicyDocument}{

(Interface and gateway endpoints) A policy to attach to the endpoint
that controls access to the service. The policy must be in valid JSON
format. If this parameter is not specified, we attach a default policy
that allows full access to the service.
}

\section{RouteTableId}{

(Gateway endpoint) One or more route table IDs.
}

\section{SubnetId}{

(Interface and Gateway Load Balancer endpoints) The ID of one or more
subnets in which to create an endpoint network interface. For a Gateway
Load Balancer endpoint, you can specify one subnet only.
}

\section{SecurityGroupId}{

(Interface endpoint) The ID of one or more security groups to associate
with the endpoint network interface.
}

\section{ClientToken}{

Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{How to Ensure Idempotency}.
}

\section{PrivateDnsEnabled}{

(Interface endpoint) Indicates whether to associate a private hosted
zone with the specified VPC. The private hosted zone contains a record
set for the default public DNS name for the service for the Region (for
example, \verb{kinesis.us-east-1.amazonaws.com}), which resolves to the
private IP addresses of the endpoint network interfaces in the VPC. This
enables you to make requests to the default public DNS name for the
service instead of the public DNS names that are automatically generated
by the VPC endpoint service.

To use a private hosted zone, you must set the following VPC attributes
to \code{true}: \code{enableDnsHostnames} and \code{enableDnsSupport}. Use
ModifyVpcAttribute to set the VPC attributes.

Default: \code{true}
}

\section{TagSpecification}{

The tags to associate with the endpoint.
}

