% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_vpc_endpoint_connection_notification}
\alias{ec2_create_vpc_endpoint_connection_notification}
\title{Create Vpc Endpoint Connection Notification}
\usage{
ec2_create_vpc_endpoint_connection_notification(
  ConnectionNotificationArn,
  ConnectionEvents,
  DryRun = NULL,
  ServiceId = NULL,
  VpcEndpointId = NULL,
  ClientToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ConnectionNotificationArn}{String. The ARN of the SNS topic for the notifications.}

\item{ConnectionEvents}{List or named list. One or more endpoint events for which to receive notifications.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{ServiceId}{String. The ID of the endpoint service. (optional)}

\item{VpcEndpointId}{String. The ID of the endpoint. (optional)}

\item{ClientToken}{String. Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a connection notification for a specified VPC endpoint or VPC
endpoint service. A connection notification notifies you of specific
endpoint events. You must create an SNS topic to receive notifications.
For more information, see \href{https://docs.aws.amazon.com/sns/latest/dg/CreateTopic.html}{Create a Topic} in
the \emph{Amazon Simple Notification Service Developer Guide}.
}
\section{ConnectionNotificationArn}{

The ARN of the SNS topic for the notifications.
}

\section{ConnectionEvents}{

One or more endpoint events for which to receive notifications. Valid
values are \code{Accept}, \code{Connect}, \code{Delete}, and \code{Reject}.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{ServiceId}{

The ID of the endpoint service.
}

\section{VpcEndpointId}{

The ID of the endpoint.
}

\section{ClientToken}{

Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{How to Ensure Idempotency}.
}

