% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_vpc_peering_connection}
\alias{ec2_create_vpc_peering_connection}
\title{Create Vpc Peering Connection}
\usage{
ec2_create_vpc_peering_connection(
  DryRun = NULL,
  PeerOwnerId = NULL,
  PeerVpcId = NULL,
  VpcId = NULL,
  PeerRegion = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{PeerOwnerId}{String. The AWS account ID of the owner of the accepter VPC.  Default: Your AWS account ID (optional)}

\item{PeerVpcId}{String. The ID of the VPC with which you are creating the VPC peering connection.(optional)}

\item{VpcId}{String. The ID of the requester VPC. You must specify this parameter in the request. (optional)}

\item{PeerRegion}{String. The Region code for the accepter VPC, if the accepter VPC is located in a Region other than...(see below)(optional)}

\item{TagSpecification}{List or named list. The tags to assign to the peering connection. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Requests a VPC peering connection between two VPCs: a requester VPC that
you own and an accepter VPC with which to create the connection. The
accepter VPC can belong to another AWS account and can be in a different
Region to the requester VPC. The requester VPC and accepter VPC cannot
have overlapping CIDR blocks.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{PeerOwnerId}{

The AWS account ID of the owner of the accepter VPC.

Default: Your AWS account ID
}

\section{PeerVpcId}{

The ID of the VPC with which you are creating the VPC peering
connection. You must specify this parameter in the request.
}

\section{VpcId}{

The ID of the requester VPC. You must specify this parameter in the
request.
}

\section{PeerRegion}{

The Region code for the accepter VPC, if the accepter VPC is located in
a Region other than the Region in which you make the request.

Default: The Region in which you make the request.
}

\section{TagSpecification}{

The tags to assign to the peering connection.
}

