% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_vpn_gateway}
\alias{ec2_create_vpn_gateway}
\title{Create Vpn Gateway}
\usage{
ec2_create_vpn_gateway(
  Type,
  AvailabilityZone = NULL,
  TagSpecification = NULL,
  AmazonSideAsn = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Type}{String. The type of VPN connection this virtual private gateway supports.}

\item{AvailabilityZone}{String. The Availability Zone for the virtual private gateway. (optional)}

\item{TagSpecification}{List or named list. The tags to apply to the virtual private gateway. (optional)}

\item{AmazonSideAsn}{Integer. A private Autonomous System Number (ASN) for the Amazon side of a BGP session.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a virtual private gateway. A virtual private gateway is the
endpoint on the VPC side of your VPN connection. You can create a
virtual private gateway before creating the VPC itself.
}
\section{Type}{

The type of VPN connection this virtual private gateway supports.
}

\section{AvailabilityZone}{

The Availability Zone for the virtual private gateway.
}

\section{TagSpecification}{

The tags to apply to the virtual private gateway.
}

\section{AmazonSideAsn}{

A private Autonomous System Number (ASN) for the Amazon side of a BGP
session. If you\'re using a 16-bit ASN, it must be in the 64512 to 65534
range. If you\'re using a 32-bit ASN, it must be in the 4200000000 to
4294967294 range.

Default: 64512
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

