% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_delete_capacity_provider}
\alias{ec2_delete_capacity_provider}
\title{Delete Capacity Provider}
\usage{
ec2_delete_capacity_provider(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Delete Capacity Provider
}
