% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_flow_logs}
\alias{ec2_delete_flow_logs}
\title{Delete Flow Logs}
\usage{
ec2_delete_flow_logs(
  FlowLogId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FlowLogId}{List or named list. One or more flow log IDs.  Constraint: Maximum of 1000 flow log IDs.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes one or more flow logs.
}
\section{FlowLogId}{

One or more flow log IDs.

Constraint: Maximum of 1000 flow log IDs.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

