% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_network_acl}
\alias{ec2_delete_network_acl}
\title{Delete Network Acl}
\usage{
ec2_delete_network_acl(
  NetworkAclId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{NetworkAclId}{String. The ID of the network ACL.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified network ACL. You can\'t delete the ACL if it\'s
associated with any subnets. You can\'t delete the default network ACL.
}
\section{NetworkAclId}{

The ID of the network ACL.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

