% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_snapshot}
\alias{ec2_delete_snapshot}
\title{Delete Snapshot}
\usage{
ec2_delete_snapshot(
  SnapshotId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{SnapshotId}{String. The ID of the EBS snapshot.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified snapshot.
}
\section{SnapshotId}{

The ID of the EBS snapshot.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

