% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_traffic_mirror_filter}
\alias{ec2_delete_traffic_mirror_filter}
\title{Delete Traffic Mirror Filter}
\usage{
ec2_delete_traffic_mirror_filter(
  TrafficMirrorFilterId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TrafficMirrorFilterId}{String. The ID of the Traffic Mirror filter.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Delete Traffic Mirror Filter
}
\section{TrafficMirrorFilterId}{

The ID of the Traffic Mirror
filter.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

