% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_traffic_mirror_filter_rule}
\alias{ec2_delete_traffic_mirror_filter_rule}
\title{Delete Traffic Mirror Filter Rule}
\usage{
ec2_delete_traffic_mirror_filter_rule(
  TrafficMirrorFilterRuleId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TrafficMirrorFilterRuleId}{String. The ID of the Traffic Mirror rule.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified Traffic Mirror rule.
}
\section{TrafficMirrorFilterRuleId}{

The ID of the Traffic Mirror rule.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

