% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_traffic_mirror_session}
\alias{ec2_delete_traffic_mirror_session}
\title{Delete Traffic Mirror Session}
\usage{
ec2_delete_traffic_mirror_session(
  TrafficMirrorSessionId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TrafficMirrorSessionId}{String. The ID of the Traffic Mirror session.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified Traffic Mirror session.
}
\section{TrafficMirrorSessionId}{

The ID of the Traffic Mirror session.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

