% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_transit_gateway_connect}
\alias{ec2_delete_transit_gateway_connect}
\title{Delete Transit Gateway Connect}
\usage{
ec2_delete_transit_gateway_connect(
  TransitGatewayAttachmentId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayAttachmentId}{String. The ID of the Connect attachment.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified Connect attachment. You must first delete any
Connect peers for the attachment.
}
\section{TransitGatewayAttachmentId}{

The ID of the Connect attachment.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

