% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_transit_gateway_connect_peer}
\alias{ec2_delete_transit_gateway_connect_peer}
\title{Delete Transit Gateway Connect Peer}
\usage{
ec2_delete_transit_gateway_connect_peer(
  TransitGatewayConnectPeerId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayConnectPeerId}{String. The ID of the Connect peer.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified Connect peer.
}
\section{TransitGatewayConnectPeerId}{

The ID of the Connect peer.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

