% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_transit_gateway_prefix_list_reference}
\alias{ec2_delete_transit_gateway_prefix_list_reference}
\title{Delete Transit Gateway Prefix List Reference}
\usage{
ec2_delete_transit_gateway_prefix_list_reference(
  TransitGatewayRouteTableId,
  PrefixListId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayRouteTableId}{String. The ID of the route table.}

\item{PrefixListId}{String. The ID of the prefix list.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes a reference (route) to a prefix list in a specified transit
gateway route table.
}
\section{TransitGatewayRouteTableId}{

The ID of the route table.
}

\section{PrefixListId}{

The ID of the prefix list.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

