% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_vpc_endpoint_connection_notifications}
\alias{ec2_delete_vpc_endpoint_connection_notifications}
\title{Delete Vpc Endpoint Connection Notifications}
\usage{
ec2_delete_vpc_endpoint_connection_notifications(
  ConnectionNotificationId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ConnectionNotificationId}{List or named list. One or more notification IDs.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes one or more VPC endpoint connection notifications.
}
\section{ConnectionNotificationId}{

One or more notification IDs.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

