% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_vpc_endpoint_service_configurations}
\alias{ec2_delete_vpc_endpoint_service_configurations}
\title{Delete Vpc Endpoint Service Configurations}
\usage{
ec2_delete_vpc_endpoint_service_configurations(
  ServiceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ServiceId}{List or named list. The IDs of one or more services.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes one or more VPC endpoint service configurations in your account.
Before you delete the endpoint service configuration, you must reject
any \code{Available} or \code{PendingAcceptance} interface endpoint connections
that are attached to the service.
}
\section{ServiceId}{

The IDs of one or more services.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

