% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_vpc_endpoints}
\alias{ec2_delete_vpc_endpoints}
\title{Delete Vpc Endpoints}
\usage{
ec2_delete_vpc_endpoints(
  VpcEndpointId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcEndpointId}{List or named list. One or more VPC endpoint IDs.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes one or more specified VPC endpoints. You can delete any of the
following types of VPC endpoints.
}
\section{VpcEndpointId}{

One or more VPC endpoint IDs.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

