% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_addresses_attribute}
\alias{ec2_describe_addresses_attribute}
\title{Describe Addresses Attribute}
\usage{
ec2_describe_addresses_attribute(
  AllocationId = NULL,
  Attribute = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AllocationId}{List or named list. [EC2-VPC] The allocation IDs. (optional)}

\item{Attribute}{String. The attribute of the IP address. (optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the attributes of the specified Elastic IP addresses. For
requirements, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/elastic-ip-addresses-eip.html#Using_Elastic_Addressing_Reverse_DNS}{Using reverse DNS for email applications}.
}
\section{AllocationId}{

[EC2-VPC] The allocation IDs.
}

\section{Attribute}{

The attribute of the IP address.
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

