% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_client_vpn_connections}
\alias{ec2_describe_client_vpn_connections}
\title{Describe Client Vpn Connections}
\usage{
ec2_describe_client_vpn_connections(
  ClientVpnEndpointId,
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{String. The ID of the Client VPN endpoint.}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes active client connections and connections that have been
terminated within the last 60 minutes for the specified Client VPN
endpoint.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN endpoint.
}

\section{Filter}{

One or more filters. Filter names and values are case-sensitive.
\itemize{
\item \code{connection-id} - The ID of the connection.
\item \code{username} - For Active Directory client authentication, the user
name of the client who established the client connection.
}
}

\section{MaxResults}{

The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the nextToken value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

