% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_client_vpn_endpoints}
\alias{ec2_describe_client_vpn_endpoints}
\title{Describe Client Vpn Endpoints}
\usage{
ec2_describe_client_vpn_endpoints(
  ClientVpnEndpointId = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filter = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{List or named list. The ID of the Client VPN endpoint. (optional)}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes one or more Client VPN endpoints in the account.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN endpoint.
}

\section{MaxResults}{

The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the nextToken value.
}

\section{Filter}{

One or more filters. Filter names and values are case-sensitive.
\itemize{
\item \code{endpoint-id} - The ID of the Client VPN endpoint.
\item \code{transport-protocol} - The transport protocol (\code{tcp} \| \code{udp}).
}
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

