% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_client_vpn_target_networks}
\alias{ec2_describe_client_vpn_target_networks}
\title{Describe Client Vpn Target Networks}
\usage{
ec2_describe_client_vpn_target_networks(
  ClientVpnEndpointId,
  AssociationIds = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filter = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{ClientVpnEndpointId}{Character. The ID of the Client VPN endpoint.}

\item{AssociationIds}{List. The IDs of the target network associations.[optional]}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.[optional]}

\item{NextToken}{Characters. The token for the next page of results[optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object (see below)[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Describes the target
networks associated with the specified Client VPN
endpoint.
}
\section{ClientVpnEndpointId}{

The ID of
the Client VPN endpoint.
}

\section{AssociationIds}{

The IDs of the target
network associations.
}

\section{MaxResults}{

The maximum number of results
to return for the request in a single page. The remaining results can be
seen by sending another request with the nextToken
value.
}

\section{Filter}{


One or more filters. Filter names and values are case-sensitive.
\itemize{
\item \code{association-id} - The ID of the association.
\item \code{target-network-id} - The ID of the subnet specified as the target
network.
\item \code{vpc-id} - The ID of the VPC in which the target network is located.
}
}

\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

