% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_coip_pools}
\alias{ec2_describe_coip_pools}
\title{Describe Coip Pools}
\usage{
ec2_describe_coip_pools(
  PoolId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{PoolId}{List or named list. The IDs of the address pools. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified customer-owned address pools or all of your
customer-owned address pools.
}
\section{PoolId}{

The IDs of the address pools.
}

\section{Filter}{

The filters. The following are the possible values:
\itemize{
\item \code{coip-pool.pool-id}
}\if{html}{\out{<div class="{=html}">}}\preformatted{
}\if{html}{\out{</div>}}
\itemize{
\item \code{coip-pool.local-gateway-route-table-id}
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

