% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_elastic_gpus}
\alias{ec2_describe_elastic_gpus}
\title{Describe Elastic Gpus}
\usage{
ec2_describe_elastic_gpus(
  ElasticGpuId = NULL,
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{ElasticGpuId}{List. The Elastic Graphics accelerator IDs.[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object (see below)[optional]}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.[optional]}

\item{NextToken}{Characters. The token for the next page of results[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Describes the Elastic Graphics accelerator
associated with your instances. For more information about Elastic
Graphics, see \href{https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/elastic-graphics.html}{Amazon Elastic Graphics}.
}
\section{ElasticGpuId}{

The Elastic Graphics accelerator
IDs.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Filter}{


The filters.
\itemize{
\item \code{availability-zone} - The Availability Zone in which the Elastic
Graphics accelerator resides.
\item \code{elastic-gpu-health} - The status of the Elastic Graphics
accelerator (\code{OK} \| \code{IMPAIRED}).
\item \code{elastic-gpu-state} - The state of the Elastic Graphics accelerator
(\code{ATTACHED}).
\item \code{elastic-gpu-type} - The type of Elastic Graphics accelerator; for
example, \code{eg1.medium}.
\item \code{instance-id} - The ID of the instance to which the Elastic Graphics
accelerator is associated.
}
}

\section{MaxResults}{

The maximum number of results to return in a single
call. To retrieve the remaining results, make another call with the
returned \code{NextToken} value. This value can be between 5 and
1000.
}

