% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_export_tasks}
\alias{ec2_describe_export_tasks}
\title{Describe Export Tasks}
\usage{
ec2_describe_export_tasks(
  ExportTaskId = NULL,
  Filter = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ExportTaskId}{List or named list. The export task IDs. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified export instance tasks or all of your export
instance tasks.
}
\section{ExportTaskId}{

The export task IDs.
}

\section{Filter}{

the filters for the export tasks.
}

