% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_fleet_instances}
\alias{ec2_describe_fleet_instances}
\title{Describe Fleet Instances}
\usage{
ec2_describe_fleet_instances(
  FleetId,
  DryRun = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Filter = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FleetId}{String. The ID of the EC2 Fleet.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the running instances for the specified EC2 Fleet.
}
\section{FleetId}{

The ID of the EC2 Fleet.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{MaxResults}{

The maximum number of results to return in a single call. Specify a
value between 1 and 1000. The default value is 1000. To retrieve the
remaining results, make another call with the returned \code{NextToken}
value.
}

\section{Filter}{

The filters.
\itemize{
\item \code{instance-type} - The instance type.
}
}

