% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_fleets}
\alias{ec2_describe_fleets}
\title{Describe Fleets}
\usage{
ec2_describe_fleets(
  DryRun = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  FleetId = NULL,
  Filter = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.[optional]}

\item{NextToken}{Characters, The token for the next page of results[optional]}

\item{FleetId}{List. The ID of the EC2 Fleets.[optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object (see below)[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Describe Fleets
}
\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{MaxResults}{

The maximum number of
results to return in a single call. Specify a value between 1 and 1000.
The default value is 1000. To retrieve the remaining results, make
another call with the returned \code{NextToken}
value.
}

\section{FleetId}{

The ID of the EC2
Fleets.
}

\section{Filter}{


The filters.
\itemize{
\item \code{activity-status} - The progress of the EC2 Fleet ( \code{error} \|
\code{pending-fulfillment} \| \code{pending-termination} \| \code{fulfilled}).
\item \code{excess-capacity-termination-policy} - Indicates whether to
terminate running instances if the target capacity is decreased
below the current EC2 Fleet size (\code{true} \| \code{false}).
\item \code{fleet-state} - The state of the EC2 Fleet (\code{submitted} \| \code{active}
\| \code{deleted} \| \code{failed} \| \code{deleted-running} \|
\code{deleted-terminating} \| \code{modifying}).
\item \code{replace-unhealthy-instances} - Indicates whether EC2 Fleet should
replace unhealthy instances (\code{true} \| \code{false}).
\item \code{type} - The type of request (\code{instant} \| \code{request} \| \code{maintain}).
}
}

