% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_host_reservations}
\alias{ec2_describe_host_reservations}
\title{Describe Host Reservations}
\usage{
ec2_describe_host_reservations(
  Filter = NULL,
  HostReservationIdSet = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object (see below)[optional]}

\item{HostReservationIdSet}{List. The host reservation IDs.[optional]}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.[optional]}

\item{NextToken}{Characters. The token for the next page of results[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Describes reservations that are associated with
Dedicated Hosts in your account.
}
\section{Filter}{


The filters.
\itemize{
\item \code{instance-family} - The instance family (for example, \code{m4}).
\item \code{payment-option} - The payment option (\code{NoUpfront} \|
\code{PartialUpfront} \| \code{AllUpfront}).
\item \code{state} - The state of the reservation (\code{payment-pending} \|
\code{payment-failed} \| \code{active} \| \code{retired}).
\item \code{tag}:\ - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}
}

\section{HostReservationIdSet}{

The host reservation IDs.
}

\section{MaxResults}{

The
maximum number of results to return for the request in a single page.
The remaining results can be seen by sending another request with the
returned \code{nextToken} value. This value can be between 5 and 500. If
\code{maxResults} is given a larger value than 500, you receive an
error.
}

