% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_iam_instance_profile_associations}
\alias{ec2_describe_iam_instance_profile_associations}
\title{Describe Iam Instance Profile Associations}
\usage{
ec2_describe_iam_instance_profile_associations(
  AssociationId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{List or named list. The IAM instance profile associations. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes your IAM instance profile associations.
}
\section{AssociationId}{

The IAM instance profile associations.
}

\section{Filter}{

The filters.
\itemize{
\item \code{instance-id} - The ID of the instance.
\item \code{state} - The state of the association (\code{associating} \|
\code{associated} \| \code{disassociating}).
}
}

\section{MaxResults}{

The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.
}

