% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_identity_id_format}
\alias{ec2_describe_identity_id_format}
\title{Describe Identity Id Format}
\usage{
ec2_describe_identity_id_format(
  PrincipalArn,
  Resource = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{PrincipalArn}{String. The ARN of the principal, which can be an IAM role, IAM user, or the root user.}

\item{Resource}{String. The type of resource: \code{bundle} \| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options}...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Describe Identity Id Format
}
\section{PrincipalArn}{

The ARN of the principal, which can be an IAM role,
IAM user, or the root user.
}

\section{Resource}{

The type of resource:
\code{bundle} \| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options} \|
\code{elastic-ip-allocation} \| \code{elastic-ip-association} \| \code{export-task} \|
\code{flow-log} \| \code{image} \| \code{import-task} \| \code{instance} \|
\code{internet-gateway} \| \code{network-acl} \| \code{network-acl-association} \|
\code{network-interface} \| \code{network-interface-attachment} \| \code{prefix-list}
\| \code{reservation} \| \code{route-table} \| \code{route-table-association} \|
\code{security-group} \| \code{snapshot} \| \code{subnet} \|
\code{subnet-cidr-block-association} \| \code{volume} \| \code{vpc} \|
\code{vpc-cidr-block-association} \| \code{vpc-endpoint} \|
\code{vpc-peering-connection} \| \code{vpn-connection} \| \code{vpn-gateway}
}

