% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_import_image_tasks}
\alias{ec2_describe_import_image_tasks}
\title{Describe Import Image Tasks}
\usage{
ec2_describe_import_image_tasks(
  DryRun = NULL,
  Filters = NULL,
  ImportTaskId = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Filters}{List or named list. Filter tasks using the \code{task-state} filter and one of the following values: \code{active}, \code{completed},...(see below)(optional)}

\item{ImportTaskId}{List or named list. The IDs of the import image tasks. (optional)}

\item{MaxResults}{Integer. The maximum number of results to return in a single call. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Displays details about an import virtual machine or import snapshot
tasks that are already created.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Filters}{

Filter tasks using the \code{task-state} filter and one of the following
values: \code{active}, \code{completed}, \code{deleting}, or \code{deleted}.
}

\section{ImportTaskId}{

The IDs of the import image tasks.
}

\section{MaxResults}{

The maximum number of results to return in a single call.
}

