% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_key_pairs}
\alias{ec2_describe_key_pairs}
\title{Describe Key Pairs}
\usage{
ec2_describe_key_pairs(
  Filter = NULL,
  KeyName = NULL,
  KeyPairId = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{KeyName}{List or named list. The key pair names.  Default: Describes all your key pairs. (optional)}

\item{KeyPairId}{List or named list. The IDs of the key pairs. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified key pairs or all of your key pairs.
}
\section{Filter}{

The filters.
\itemize{
\item \code{key-pair-id} - The ID of the key pair.
\item \code{fingerprint} - The fingerprint of the key pair.
\item \code{key-name} - The name of the key pair.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
\item \code{tag}:\ - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
}
}

\section{KeyName}{

The key pair names.

Default: Describes all your key pairs.
}

\section{KeyPairId}{

The IDs of the key pairs.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

