% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_local_gateway_route_table_virtual_interface_group_associations}
\alias{ec2_describe_local_gateway_route_table_virtual_interface_group_associations}
\title{Describe Local Gateway Route Table Virtual Interface Group Associations}
\usage{
ec2_describe_local_gateway_route_table_virtual_interface_group_associations(
  LocalGatewayRouteTableVirtualInterfaceGroupAssociationId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{LocalGatewayRouteTableVirtualInterfaceGroupAssociationId}{List or named list. The IDs of the associations. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the associations between virtual interface groups and local
gateway route tables.
}
\section{LocalGatewayRouteTableVirtualInterfaceGroupAssociationId}{

The IDs of the associations.
}

\section{Filter}{

One or more filters.
\itemize{
\item \code{local-gateway-id} - The ID of a local gateway.
\item \code{local-gateway-route-table-id} - The ID of the local gateway route
table.
\item \code{local-gateway-route-table-virtual-interface-group-association-id} -
The ID of the association.
\item \code{local-gateway-route-table-virtual-interface-group-id} - The ID of
the virtual interface group.
\item \code{state} - The state of the association.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

