% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_scheduled_instance_availability}
\alias{ec2_describe_scheduled_instance_availability}
\title{Describe Scheduled Instance Availability}
\usage{
ec2_describe_scheduled_instance_availability(
  FirstSlotStartTimeRange,
  Recurrence,
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  MaxSlotDurationInHours = NULL,
  MinSlotDurationInHours = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FirstSlotStartTimeRange}{Object. The time period for the first schedule to start.}

\item{Recurrence}{Object. The schedule recurrence.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object. (see below)[optional]}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.[optional]}

\item{MaxSlotDurationInHours}{Integer. The maximum available duration, in hours.[optional]}

\item{MinSlotDurationInHours}{Integer. The minimum available duration, in hours.[optional]}

\item{NextToken}{Characters, The token for the next page of results.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describe Scheduled Instance Availability
}
\section{FirstSlotStartTimeRange}{

The time period
for the first schedule to start.
}

\section{Recurrence}{

The schedule
recurrence.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Filter}{


The filters.
\itemize{
\item \code{availability-zone} - The Availability Zone (for example,
\verb{us-west-2a}).
\item \code{instance-type} - The instance type (for example, \code{c4.large}).
\item \code{network-platform} - The network platform (\code{EC2-Classic} or
\code{EC2-VPC}).
\item \code{platform} - The platform (\code{Linux/UNIX} or \code{Windows}).
}
}

\section{MaxResults}{

The maximum number of results to return in a single
call. This value can be between 5 and 300. The default value is 300. To
retrieve the remaining results, make another call with the returned
\code{NextToken} value.
}

\section{MaxSlotDurationInHours}{

The maximum available duration,
in hours. This value must be greater than \code{MinSlotDurationInHours} and
less than 1,720.
}

\section{MinSlotDurationInHours}{

The minimum available duration, in
hours. The minimum required duration is 1,200 hours per year. For
example, the minimum daily schedule is 4 hours, the minimum weekly
schedule is 24 hours, and the minimum monthly schedule is 100
hours.
}

