% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_transit_gateway_multicast_domains}
\alias{ec2_describe_transit_gateway_multicast_domains}
\title{Describe Transit Gateway Multicast Domains}
\usage{
ec2_describe_transit_gateway_multicast_domains(
  TransitGatewayMulticastDomainIds = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayMulticastDomainIds}{Array. The ID of the transit gateway multicast domain.\link{optional}}

\item{Filter}{No description can be found.}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.\link{optional}}

\item{NextToken}{Characters, The token for the next page of results.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes one or more transit gateway
multicast domains.
}
\section{TransitGatewayMulticastDomainIds}{

The ID of the transit
gateway multicast domain.
}

\section{Filter}{


One or more filters. The possible values are:
\itemize{
\item \code{state} - The state of the transit gateway multicast domain. Valid
values are \code{pending} \| \code{available} \| \code{deleting} \| \code{deleted}.
\item \code{transit-gateway-id} - The ID of the transit gateway.
\item \code{transit-gateway-multicast-domain-id} - The ID of the transit
gateway multicast domain.
}
}

\section{MaxResults}{

The maximum number of results to return with a
single call. To retrieve the remaining results, make another call with
the returned \code{nextToken} value.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

