% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_volume_attribute}
\alias{ec2_describe_volume_attribute}
\title{Describe Volume Attribute}
\usage{
ec2_describe_volume_attribute(
  Attribute,
  VolumeId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The attribute of the volume. This parameter is required.}

\item{VolumeId}{String. The ID of the volume.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describe Volume Attribute
}
\section{Attribute}{

The attribute of the
volume. This parameter is required.
}

\section{VolumeId}{

The ID of the
volume.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

