% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_volumes_modifications}
\alias{ec2_describe_volumes_modifications}
\title{Describe Volumes Modifications}
\usage{
ec2_describe_volumes_modifications(
  DryRun = NULL,
  VolumeId = NULL,
  Filter = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{VolumeId}{List or named list. The IDs of the volumes. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results (up to a limit of 500) to be returned in a paginated request. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the most recent volume modification request for the specified
EBS volumes.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{VolumeId}{

The IDs of the volumes.
}

\section{Filter}{

The filters.
\itemize{
\item \code{modification-state} - The current modification state (modifying \|
optimizing \| completed \| failed).
\item \code{original-iops} - The original IOPS rate of the volume.
\item \code{original-size} - The original size of the volume, in GiB.
\item \code{original-volume-type} - The original volume type of the volume
(standard \| io1 \| io2 \| gp2 \| sc1 \| st1).
\item \code{originalMultiAttachEnabled} - Indicates whether Multi-Attach
support was enabled (true \| false).
\item \code{start-time} - The modification start time.
\item \code{target-iops} - The target IOPS rate of the volume.
\item \code{target-size} - The target size of the volume, in GiB.
\item \code{target-volume-type} - The target volume type of the volume
(standard \| io1 \| io2 \| gp2 \| sc1 \| st1).
\item \code{targetMultiAttachEnabled} - Indicates whether Multi-Attach support
is to be enabled (true \| false).
\item \code{volume-id} - The ID of the volume.
}
}

\section{MaxResults}{

The maximum number of results (up to a limit of 500) to be returned in a
paginated request.
}

