% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_detach_internet_gateway}
\alias{ec2_detach_internet_gateway}
\title{Detach Internet Gateway}
\usage{
ec2_detach_internet_gateway(
  InternetGatewayId,
  VpcId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InternetGatewayId}{String. The ID of the internet gateway.}

\item{VpcId}{String. The ID of the VPC.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Detaches an internet gateway from a VPC, disabling connectivity between
the internet and the VPC. The VPC must not contain any running instances
with Elastic IP addresses or public IPv4 addresses.
}
\section{InternetGatewayId}{

The ID of the internet gateway.
}

\section{VpcId}{

The ID of the VPC.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

