% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disable_vgw_route_propagation}
\alias{ec2_disable_vgw_route_propagation}
\title{Disable Vgw Route Propagation}
\usage{
ec2_disable_vgw_route_propagation(
  GatewayId,
  RouteTableId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{GatewayId}{String. The ID of the virtual private gateway.}

\item{RouteTableId}{String. The ID of the route table.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Disables a virtual private gateway (VGW) from propagating routes to a
specified route table of a VPC.
}
\section{GatewayId}{

The ID of the virtual private gateway.
}

\section{RouteTableId}{

The ID of the route table.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

