% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disassociate_iam_instance_profile}
\alias{ec2_disassociate_iam_instance_profile}
\title{Disassociate Iam Instance Profile}
\usage{
ec2_disassociate_iam_instance_profile(
  AssociationId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. The ID of the IAM instance profile association.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Disassociates an IAM instance profile from a running or stopped
instance.
}
\section{AssociationId}{

The ID of the IAM instance profile association.
}

