% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_enable_fast_snapshot_restores}
\alias{ec2_enable_fast_snapshot_restores}
\title{Enable Fast Snapshot Restores}
\usage{
ec2_enable_fast_snapshot_restores(
  AvailabilityZone,
  SourceSnapshotId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AvailabilityZone}{List or named list. One or more Availability Zones. For example, \verb{us-east-2a}.}

\item{SourceSnapshotId}{List or named list. The IDs of one or more snapshots.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Enables fast snapshot restores for the specified snapshots in the
specified Availability Zones.
}
\section{AvailabilityZone}{

One or more Availability Zones. For example, \verb{us-east-2a}.
}

\section{SourceSnapshotId}{

The IDs of one or more snapshots. For example, \verb{snap-1234567890abcdef0}.
You can specify a snapshot that was shared with you from another AWS
account.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

