% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_export_client_vpn_client_configuration}
\alias{ec2_export_client_vpn_client_configuration}
\title{Export Client Vpn Client Configuration}
\usage{
ec2_export_client_vpn_client_configuration(
  ClientVpnEndpointId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{String. The ID of the Client VPN endpoint.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Downloads the
contents of the Client VPN endpoint configuration file for the specified
Client VPN endpoint. The Client VPN endpoint configuration file includes
the Client VPN endpoint and certificate information clients need to
establish a connection with the Client VPN
endpoint.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN
endpoint.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

