% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_console_output}
\alias{ec2_get_console_output}
\title{Get Console Output}
\usage{
ec2_get_console_output(
  InstanceId,
  DryRun = NULL,
  Latest = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{String. The ID of the instance.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Latest}{Boolean. When enabled, retrieves the latest console output for the instance.  Default: disabled (\code{false}) (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Gets the console output for the specified instance. For Linux instances,
the instance console output displays the exact console output that would
normally be displayed on a physical monitor attached to a computer. For
Windows instances, the instance console output includes the last three
system event log errors.
}
\section{InstanceId}{

The ID of the instance.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Latest}{

When enabled, retrieves the latest console output for the instance.

Default: disabled (\code{false})
}

