% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_console_screenshot}
\alias{ec2_get_console_screenshot}
\title{Get Console Screenshot}
\usage{
ec2_get_console_screenshot(
  InstanceId,
  DryRun = NULL,
  WakeUp = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{String. The ID of the instance.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{WakeUp}{Boolean. When set to \code{true}, acts as keystroke input and wakes up an instance that\'s in standby or...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Retrieve a JPG-format screenshot of a running instance to help with
troubleshooting.
}
\section{InstanceId}{

The ID of the instance.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{WakeUp}{

When set to \code{true}, acts as keystroke input and wakes up an instance
that\'s in standby or \'sleep\' mode.
}

