% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_ebs_default_kms_key_id}
\alias{ec2_get_ebs_default_kms_key_id}
\title{Get Ebs Default Kms Key Id}
\usage{
ec2_get_ebs_default_kms_key_id(DryRun = NULL, simplify = TRUE, others = list())
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the default customer master key (CMK) for EBS encryption by
default for your account in this Region. You can change the default CMK
for encryption by default using ModifyEbsDefaultKmsKeyId or
ResetEbsDefaultKmsKeyId.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

