% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_host_reservation_purchase_preview}
\alias{ec2_get_host_reservation_purchase_preview}
\title{Get Host Reservation Purchase Preview}
\usage{
ec2_get_host_reservation_purchase_preview(
  HostIdSet,
  OfferingId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{HostIdSet}{List. The IDs of the Dedicated Hosts with which the reservation is associated.}

\item{OfferingId}{String. The offering ID of the reservation.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Get Host Reservation Purchase Preview
}
\section{HostIdSet}{

The IDs of
the Dedicated Hosts with which the reservation is
associated.
}

\section{OfferingId}{

The offering ID of the
reservation.
}

