% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_reserved_instances_exchange_quote}
\alias{ec2_get_reserved_instances_exchange_quote}
\title{Get Reserved Instances Exchange Quote}
\usage{
ec2_get_reserved_instances_exchange_quote(
  ReservedInstanceId,
  DryRun = NULL,
  TargetConfiguration = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ReservedInstanceId}{List or named list. The IDs of the Convertible Reserved Instances to exchange.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{TargetConfiguration}{List or named list. The configuration of the target Convertible Reserved Instance to exchange for your current...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Returns a quote and exchange information for exchanging one or more
specified Convertible Reserved Instances for a new Convertible Reserved
Instance. If the exchange cannot be performed, the reason is returned in
the response. Use AcceptReservedInstancesExchangeQuote to perform the
exchange.
}
\section{ReservedInstanceId}{

The IDs of the Convertible Reserved Instances to exchange.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{TargetConfiguration}{

The configuration of the target Convertible Reserved Instance to
exchange for your current Convertible Reserved Instances.
}

