% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_transit_gateway_attachment_propagations}
\alias{ec2_get_transit_gateway_attachment_propagations}
\title{Get Transit Gateway Attachment Propagations}
\usage{
ec2_get_transit_gateway_attachment_propagations(
  TransitGatewayAttachmentId,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayAttachmentId}{String. The ID of the attachment.}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Lists the route tables to which the specified resource attachment
propagates routes.
}
\section{TransitGatewayAttachmentId}{

The ID of the attachment.
}

\section{Filter}{

One or more filters. The possible values are:
\itemize{
\item \code{transit-gateway-route-table-id} - The ID of the transit gateway
route table.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

