% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_import_instance}
\alias{ec2_import_instance}
\title{Import Instance}
\usage{
ec2_import_instance(
  Platform,
  Description = NULL,
  DiskImage = NULL,
  DryRun = NULL,
  LaunchSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Platform}{String. The instance operating system.}

\item{Description}{String. A description for the instance being imported. (optional)}

\item{DiskImage}{List or named list. The disk image. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{LaunchSpecification}{Object. The launch specification. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates an import instance task using metadata from the specified disk
image. \code{ImportInstance} only supports single-volume VMs. To import
multi-volume VMs, use ImportImage. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/CommandLineReference/ec2-cli-vmimport-export.html}{Importing a Virtual Machine Using the Amazon EC2 CLI}.
}
\section{Platform}{

The instance operating system.
}

\section{Description}{

A description for the instance being imported.
}

\section{DiskImage}{

The disk image.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{LaunchSpecification}{

The launch specification.
}

