% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_list_account_settings}
\alias{ec2_list_account_settings}
\title{List Account Settings}
\usage{
ec2_list_account_settings(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Lists the account settings for a specified
principal.
}
\section{maxResults}{

Pagination limit
}

