% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_list_container_instances}
\alias{ec2_list_container_instances}
\title{List Container Instances}
\usage{
ec2_list_container_instances(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Returns a list of
container instances in a specified cluster. You can filter the results
of a \code{ListContainerInstances} operation with cluster query language
statements inside the \code{filter} parameter. For more information, see
\href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/cluster-query-language.html}{Cluster Query Language}
in the \emph{Amazon Elastic Container Service Developer
Guide}.
}
\section{maxResults}{

Pagination
limit
}

