% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_list_task_definition_families}
\alias{ec2_list_task_definition_families}
\title{List Task Definition Families}
\usage{
ec2_list_task_definition_families(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
List Task Definition Families
}
\section{maxResults}{

Pagination
limit
}

