% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_capacity_reservation}
\alias{ec2_modify_capacity_reservation}
\title{Modify Capacity Reservation}
\usage{
ec2_modify_capacity_reservation(
  CapacityReservationId,
  InstanceCount = NULL,
  EndDate = NULL,
  EndDateType = NULL,
  Accept = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{CapacityReservationId}{String. The ID of the Capacity Reservation.}

\item{InstanceCount}{Integer. The number of instances for which to reserve capacity.\link{optional}}

\item{EndDate}{String.   The date and time at which the Capacity Reservation expires.\link{optional}}

\item{EndDateType}{String.   Indicates the way in which the Capacity Reservation ends.\link{optional}}

\item{Accept}{Boolean. Reserved. Capacity Reservations you have created are accepted by default.\link{optional}}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Modifies a Capacity Reservation\'s capacity and the
conditions under which it is to be released. You cannot change a
Capacity Reservation\'s instance type, EBS optimization, instance store
settings, platform, Availability Zone, or instance eligibility. If you
need to modify any of these attributes, we recommend that you cancel the
Capacity Reservation, and then create a new one with the required
attributes.
}
\section{CapacityReservationId}{

The ID of the Capacity
Reservation.
}

\section{InstanceCount}{

The number of instances for which to
reserve capacity.
}

\section{EndDate}{


The date and time at which the Capacity Reservation expires. When a
Capacity Reservation expires, the reserved capacity is released and you
can no longer launch instances into it. The Capacity Reservation\'s
state changes to \code{expired} when it reaches its end date and time.

The Capacity Reservation is cancelled within an hour from the specified
time. For example, if you specify 5/31/2019, 13:30:55, the Capacity
Reservation is guaranteed to end between 13:30:55 and 14:30:55 on
5/31/2019.

You must provide an \code{EndDate} value if \code{EndDateType} is \code{limited}. Omit
\code{EndDate} if \code{EndDateType} is \code{unlimited}.
}

\section{EndDateType}{


Indicates the way in which the Capacity Reservation ends. A Capacity
Reservation can have one of the following end types:
\itemize{
\item \code{unlimited} - The Capacity Reservation remains active until you
explicitly cancel it. Do not provide an \code{EndDate} value if
\code{EndDateType} is \code{unlimited}.
\item \code{limited} - The Capacity Reservation expires automatically at a
specified date and time. You must provide an \code{EndDate} value if
\code{EndDateType} is \code{limited}.
}
}

\section{Accept}{

Reserved. Capacity Reservations you have created
are accepted by default.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

