% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_default_credit_specification}
\alias{ec2_modify_default_credit_specification}
\title{Modify Default Credit Specification}
\usage{
ec2_modify_default_credit_specification(
  InstanceFamily,
  CpuCredits,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceFamily}{String. The instance family.}

\item{CpuCredits}{String. The credit option for CPU usage of the instance family.  Valid Values: \code{standard} \| \code{unlimited}}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the default credit option for CPU usage of burstable
performance instances. The default credit option is set at the account
level per AWS Region, and is specified per instance family. All new
burstable performance instances in the account launch using the default
credit option.
}
\section{InstanceFamily}{

The instance family.
}

\section{CpuCredits}{

The credit option for CPU usage of the instance family.

Valid Values: \code{standard} \| \code{unlimited}
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

