% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_fleet}
\alias{ec2_modify_fleet}
\title{Modify Fleet}
\usage{
ec2_modify_fleet(
  FleetId,
  DryRun = NULL,
  ExcessCapacityTerminationPolicy = NULL,
  LaunchTemplateConfig = NULL,
  TargetCapacitySpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FleetId}{String. The ID of the EC2 Fleet.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{ExcessCapacityTerminationPolicy}{String. Indicates whether running instances should be terminated if the total target capacity of the...(see below)(optional)}

\item{LaunchTemplateConfig}{List or named list. The launch template and overrides. (optional)}

\item{TargetCapacitySpecification}{Object. The size of the EC2 Fleet. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the specified EC2 Fleet.
}
\section{FleetId}{

The ID of the EC2 Fleet.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{ExcessCapacityTerminationPolicy}{

Indicates whether running instances should be terminated if the total
target capacity of the EC2 Fleet is decreased below the current size of
the EC2 Fleet.
}

\section{LaunchTemplateConfig}{

The launch template and overrides.
}

\section{TargetCapacitySpecification}{

The size of the EC2 Fleet.
}

