% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_id_format}
\alias{ec2_modify_id_format}
\title{Modify Id Format}
\usage{
ec2_modify_id_format(Resource, UseLongIds, simplify = TRUE, others = list())
}
\arguments{
\item{Resource}{String. The type of resource: \code{bundle} \| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options}...(see below)}

\item{UseLongIds}{Boolean. Indicate whether the resource should use longer IDs (17-character IDs).}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the ID format for the specified resource on a per-Region basis.
You can specify that resources should receive longer IDs (17-character
IDs) when they are created.
}
\section{Resource}{

The type of resource: \code{bundle} \| \code{conversion-task} \|
\code{customer-gateway} \| \code{dhcp-options} \| \code{elastic-ip-allocation} \|
\code{elastic-ip-association} \| \code{export-task} \| \code{flow-log} \| \code{image} \|
\code{import-task} \| \code{internet-gateway} \| \code{network-acl} \|
\code{network-acl-association} \| \code{network-interface} \|
\code{network-interface-attachment} \| \code{prefix-list} \| \code{route-table} \|
\code{route-table-association} \| \code{security-group} \| \code{subnet} \|
\code{subnet-cidr-block-association} \| \code{vpc} \| \code{vpc-cidr-block-association}
\| \code{vpc-endpoint} \| \code{vpc-peering-connection} \| \code{vpn-connection} \|
\code{vpn-gateway}.

Alternatively, use the \code{all-current} option to include all resource
types that are currently within their opt-in period for longer IDs.
}

\section{UseLongIds}{

Indicate whether the resource should use longer IDs (17-character IDs).
}

