% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_image_attribute}
\alias{ec2_modify_image_attribute}
\title{Modify Image Attribute}
\usage{
ec2_modify_image_attribute(
  ImageId,
  Attribute = NULL,
  Description = NULL,
  LaunchPermission = NULL,
  OperationType = NULL,
  ProductCode = NULL,
  UserGroup = NULL,
  UserId = NULL,
  Value = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ImageId}{String. The ID of the AMI.}

\item{Attribute}{String. The name of the attribute to modify.(optional)}

\item{Description}{Object. A new description for the AMI. (optional)}

\item{LaunchPermission}{Object. A new launch permission for the AMI. (optional)}

\item{OperationType}{String. The operation type.(optional)}

\item{ProductCode}{List or named list. The DevPay product codes. After you add a product code to an AMI, it can\'t be removed. (optional)}

\item{UserGroup}{List or named list. The user groups.(optional)}

\item{UserId}{List or named list. The AWS account IDs.(optional)}

\item{Value}{String. The value of the attribute being modified.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the specified attribute of the specified AMI. You can specify
only one attribute at a time. You can use the \code{Attribute} parameter to
specify the attribute or one of the following parameters: \code{Description},
\code{LaunchPermission}, or \code{ProductCode}.
}
\section{ImageId}{

The ID of the AMI.
}

\section{Attribute}{

The name of the attribute to modify. The valid values are \code{description},
\code{launchPermission}, and \code{productCodes}.
}

\section{Description}{

A new description for the AMI.
}

\section{LaunchPermission}{

A new launch permission for the AMI.
}

\section{OperationType}{

The operation type. This parameter can be used only when the \code{Attribute}
parameter is \code{launchPermission}.
}

\section{ProductCode}{

The DevPay product codes. After you add a product code to an AMI, it
can\'t be removed.
}

\section{UserGroup}{

The user groups. This parameter can be used only when the \code{Attribute}
parameter is \code{launchPermission}.
}

\section{UserId}{

The AWS account IDs. This parameter can be used only when the
\code{Attribute} parameter is \code{launchPermission}.
}

\section{Value}{

The value of the attribute being modified. This parameter can be used
only when the \code{Attribute} parameter is \code{description} or \code{productCodes}.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

