% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_managed_prefix_list}
\alias{ec2_modify_managed_prefix_list}
\title{Modify Managed Prefix List}
\usage{
ec2_modify_managed_prefix_list(
  PrefixListId,
  DryRun = NULL,
  CurrentVersion = NULL,
  PrefixListName = NULL,
  AddEntry = NULL,
  RemoveEntry = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{PrefixListId}{String. The ID of the prefix list.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{CurrentVersion}{Integer. The current version of the prefix list. (optional)}

\item{PrefixListName}{String. A name for the prefix list. (optional)}

\item{AddEntry}{List or named list. One or more entries to add to the prefix list. (optional)}

\item{RemoveEntry}{List or named list. One or more entries to remove from the prefix list. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the specified managed prefix list.
}
\section{PrefixListId}{

The ID of the prefix list.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{CurrentVersion}{

The current version of the prefix list.
}

\section{PrefixListName}{

A name for the prefix list.
}

\section{AddEntry}{

One or more entries to add to the prefix list.
}

\section{RemoveEntry}{

One or more entries to remove from the prefix list.
}

