% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_reserved_instances}
\alias{ec2_modify_reserved_instances}
\title{Modify Reserved Instances}
\usage{
ec2_modify_reserved_instances(
  ReservedInstancesId,
  ReservedInstancesConfigurationSetItemType,
  ClientToken = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout(),
  region = aws_get_region()
)
}
\arguments{
\item{ReservedInstancesId}{List. The IDs of the Reserved Instances to modify.}

\item{ReservedInstancesConfigurationSetItemType}{List. The configuration settings for the Reserved Instances to modify.}

\item{ClientToken}{Character. A unique, case-sensitive token you provide to ensure idempotency of your modification request.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}

\item{region}{Character. The region of the AWS service.}
}
\value{
A list object or a character vector
}
\description{
Modify Reserved Instances
}
\section{ReservedInstancesId}{

The IDs of the Reserved Instances to
modify.
}

\section{ReservedInstancesConfigurationSetItemType}{

The configuration settings for the Reserved
Instances to modify.
}

\section{ClientToken}{

A unique, case-sensitive token
you provide to ensure idempotency of your modification request. For more
information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.
}

