% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_spot_fleet_request}
\alias{ec2_modify_spot_fleet_request}
\title{Modify Spot Fleet Request}
\usage{
ec2_modify_spot_fleet_request(
  SpotFleetRequestId,
  ExcessCapacityTerminationPolicy = NULL,
  LaunchTemplateConfig = NULL,
  TargetCapacity = NULL,
  OnDemandTargetCapacity = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{SpotFleetRequestId}{String. The ID of the Spot Fleet request.}

\item{ExcessCapacityTerminationPolicy}{String. Indicates whether running Spot Instances should be terminated if the target capacity of the...(see below)(optional)}

\item{LaunchTemplateConfig}{List or named list. The launch template and overrides.(optional)}

\item{TargetCapacity}{Integer. The size of the fleet. (optional)}

\item{OnDemandTargetCapacity}{Integer. The number of On-Demand Instances in the fleet. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the specified Spot Fleet request.
}
\section{SpotFleetRequestId}{

The ID of the Spot Fleet request.
}

\section{ExcessCapacityTerminationPolicy}{

Indicates whether running Spot Instances should be terminated if the
target capacity of the Spot Fleet request is decreased below the current
size of the Spot Fleet.
}

\section{LaunchTemplateConfig}{

The launch template and overrides. You can only use this parameter if
you specified a launch template (\code{LaunchTemplateConfigs}) in your Spot
Fleet request. If you specified \code{LaunchSpecifications} in your Spot
Fleet request, then omit this parameter.
}

\section{TargetCapacity}{

The size of the fleet.
}

\section{OnDemandTargetCapacity}{

The number of On-Demand Instances in the fleet.
}

