% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_transit_gateway_prefix_list_reference}
\alias{ec2_modify_transit_gateway_prefix_list_reference}
\title{Modify Transit Gateway Prefix List Reference}
\usage{
ec2_modify_transit_gateway_prefix_list_reference(
  TransitGatewayRouteTableId,
  PrefixListId,
  TransitGatewayAttachmentId = NULL,
  Blackhole = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayRouteTableId}{String. The ID of the transit gateway route table.}

\item{PrefixListId}{String. The ID of the prefix list.}

\item{TransitGatewayAttachmentId}{String. The ID of the attachment to which traffic is routed. (optional)}

\item{Blackhole}{Boolean. Indicates whether to drop traffic that matches this route. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies a reference (route) to a prefix list in a specified transit
gateway route table.
}
\section{TransitGatewayRouteTableId}{

The ID of the transit gateway route table.
}

\section{PrefixListId}{

The ID of the prefix list.
}

\section{TransitGatewayAttachmentId}{

The ID of the attachment to which traffic is routed.
}

\section{Blackhole}{

Indicates whether to drop traffic that matches this route.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

