% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_vpc_endpoint_service_permissions}
\alias{ec2_modify_vpc_endpoint_service_permissions}
\title{Modify Vpc Endpoint Service Permissions}
\usage{
ec2_modify_vpc_endpoint_service_permissions(
  ServiceId,
  DryRun = NULL,
  AddAllowedPrincipals = NULL,
  RemoveAllowedPrincipals = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ServiceId}{String. The ID of the service.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{AddAllowedPrincipals}{List or named list. The Amazon Resource Names (ARN) of one or more principals.(optional)}

\item{RemoveAllowedPrincipals}{List or named list. The Amazon Resource Names (ARN) of one or more principals.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the permissions for your \href{https://docs.aws.amazon.com/vpc/latest/userguide/endpoint-service.html}{VPC endpoint service}.
You can add or remove permissions for service consumers (IAM users, IAM
roles, and AWS accounts) to connect to your endpoint service.
}
\section{ServiceId}{

The ID of the service.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{AddAllowedPrincipals}{

The Amazon Resource Names (ARN) of one or more principals. Permissions
are granted to the principals in this list. To grant permissions to all
principals, specify an asterisk (\*).
}

\section{RemoveAllowedPrincipals}{

The Amazon Resource Names (ARN) of one or more principals. Permissions
are revoked for principals in this list.
}

