% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_vpc_peering_connection_options}
\alias{ec2_modify_vpc_peering_connection_options}
\title{Modify Vpc Peering Connection Options}
\usage{
ec2_modify_vpc_peering_connection_options(
  VpcPeeringConnectionId,
  AccepterPeeringConnectionOptions = NULL,
  DryRun = NULL,
  RequesterPeeringConnectionOptions = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcPeeringConnectionId}{String. The ID of the VPC peering connection.}

\item{AccepterPeeringConnectionOptions}{Object. The VPC peering connection options for the accepter VPC. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{RequesterPeeringConnectionOptions}{Object. The VPC peering connection options for the requester VPC. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the VPC peering connection options on one side of a VPC peering
connection. You can do the following:
}
\section{VpcPeeringConnectionId}{

The ID of the VPC peering connection.
}

\section{AccepterPeeringConnectionOptions}{

The VPC peering connection options for the accepter VPC.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{RequesterPeeringConnectionOptions}{

The VPC peering connection options for the requester VPC.
}

