% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reboot_instances}
\alias{ec2_reboot_instances}
\title{Reboot Instances}
\usage{
ec2_reboot_instances(
  InstanceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{Array. The instance IDs.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Reboot Instances
}
\section{InstanceId}{

The
instance IDs.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

