% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_register_instance_event_notification_attributes}
\alias{ec2_register_instance_event_notification_attributes}
\title{Register Instance Event Notification Attributes}
\usage{
ec2_register_instance_event_notification_attributes(
  DryRun = NULL,
  InstanceTagAttribute = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{InstanceTagAttribute}{Object. Information about the tag keys to register. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Registers a set of tag keys to include in scheduled event notifications
for your resources.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{InstanceTagAttribute}{

Information about the tag keys to register.
}

