% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_release_hosts}
\alias{ec2_release_hosts}
\title{Release Hosts}
\usage{
ec2_release_hosts(HostId, simplify = TRUE, others = list())
}
\arguments{
\item{HostId}{List or named list. The IDs of the Dedicated Hosts to release.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
When you no longer want to use an On-Demand Dedicated Host it can be
released. On-Demand billing is stopped and the host goes into \code{released}
state. The host ID of Dedicated Hosts that have been released can no
longer be specified in another request, for example, to modify the host.
You must stop or terminate all instances on a host before it can be
released.
}
\section{HostId}{

The IDs of the Dedicated Hosts to release.
}

