% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_replace_network_acl_association}
\alias{ec2_replace_network_acl_association}
\title{Replace Network Acl Association}
\usage{
ec2_replace_network_acl_association(
  AssociationId,
  NetworkAclId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. The ID of the current association between the original network ACL and the subnet.}

\item{NetworkAclId}{String. The ID of the new network ACL to associate with the subnet.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Changes which network ACL a subnet is associated with. By default when
you create a subnet, it\'s automatically associated with the default
network ACL. For more information, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/VPC_ACLs.html}{Network ACLs} in
the \emph{Amazon Virtual Private Cloud User Guide}.
}
\section{AssociationId}{

The ID of the current association between the original network ACL and
the subnet.
}

\section{NetworkAclId}{

The ID of the new network ACL to associate with the subnet.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

