% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_replace_transit_gateway_route}
\alias{ec2_replace_transit_gateway_route}
\title{Replace Transit Gateway Route}
\usage{
ec2_replace_transit_gateway_route(
  DestinationCidrBlock,
  TransitGatewayRouteTableId,
  TransitGatewayAttachmentId = NULL,
  Blackhole = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DestinationCidrBlock}{String. The CIDR range used for the destination match.}

\item{TransitGatewayRouteTableId}{String. The ID of the route table.}

\item{TransitGatewayAttachmentId}{String. The ID of the attachment. (optional)}

\item{Blackhole}{Boolean. Indicates whether traffic matching this route is to be dropped. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Replaces the specified route in the specified transit gateway route
table.
}
\section{DestinationCidrBlock}{

The CIDR range used for the destination match. Routing decisions are
based on the most specific match.
}

\section{TransitGatewayRouteTableId}{

The ID of the route table.
}

\section{TransitGatewayAttachmentId}{

The ID of the attachment.
}

\section{Blackhole}{

Indicates whether traffic matching this route is to be dropped.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

