% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reset_ebs_default_kms_key_id}
\alias{ec2_reset_ebs_default_kms_key_id}
\title{Reset Ebs Default Kms Key Id}
\usage{
ec2_reset_ebs_default_kms_key_id(
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Resets the default customer master key (CMK) for EBS encryption for your
account in this Region to the AWS managed CMK for EBS.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

