% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reset_fpga_image_attribute}
\alias{ec2_reset_fpga_image_attribute}
\title{Reset Fpga Image Attribute}
\usage{
ec2_reset_fpga_image_attribute(
  FpgaImageId,
  DryRun = NULL,
  Attribute = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FpgaImageId}{String. The ID of the AFI.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{Attribute}{String. The attribute.\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Resets the specified attribute of the specified
Amazon FPGA Image (AFI) to its default value. You can only reset the
load permission attribute.
}
\section{FpgaImageId}{

The ID of the
AFI.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Attribute}{

The
attribute.
}

