% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reset_image_attribute}
\alias{ec2_reset_image_attribute}
\title{Reset Image Attribute}
\usage{
ec2_reset_image_attribute(
  Attribute,
  ImageId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The attribute to reset (currently you can only reset the launch permission attribute).}

\item{ImageId}{String. The ID of the AMI.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Resets an attribute of an AMI to its default value.
}
\section{Attribute}{

The attribute to reset (currently you can only reset the launch
permission attribute).
}

\section{ImageId}{

The ID of the AMI.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

