% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reset_instance_attribute}
\alias{ec2_reset_instance_attribute}
\title{Reset Instance Attribute}
\usage{
ec2_reset_instance_attribute(
  Attribute,
  InstanceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The attribute to reset.}

\item{InstanceId}{String. The ID of the instance.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Resets an attribute of an instance to its default value. To reset the
\code{kernel} or \code{ramdisk}, the instance must be in a stopped state. To reset
the \code{sourceDestCheck}, the instance can be either running or stopped.
}
\section{Attribute}{

The attribute to reset.

You can only reset the following attributes: \code{kernel} \| \code{ramdisk} \|
\code{sourceDestCheck}. To change an instance attribute, use
ModifyInstanceAttribute.
}

\section{InstanceId}{

The ID of the instance.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

