% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_reset_snapshot_attribute}
\alias{ec2_reset_snapshot_attribute}
\title{Reset Snapshot Attribute}
\usage{
ec2_reset_snapshot_attribute(
  Attribute,
  SnapshotId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The attribute to reset.}

\item{SnapshotId}{String. The ID of the snapshot.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Resets permission settings for the specified snapshot.
}
\section{Attribute}{

The attribute to reset. Currently, only the attribute for permission to
create volumes can be reset.
}

\section{SnapshotId}{

The ID of the snapshot.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

