% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_revoke_security_group_egress}
\alias{ec2_revoke_security_group_egress}
\title{Revoke Security Group Egress}
\usage{
ec2_revoke_security_group_egress(
  GroupId,
  DryRun = NULL,
  IpPermissions = NULL,
  CidrIp = NULL,
  FromPort = NULL,
  IpProtocol = NULL,
  ToPort = NULL,
  SourceSecurityGroupName = NULL,
  SourceSecurityGroupOwnerId = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{GroupId}{Character. The ID of the security group.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{IpPermissions}{List. The sets of IP permissions.[optional]}

\item{CidrIp}{Character. Not supported. Use a set of IP permissions to specify the CIDR.[optional]}

\item{FromPort}{Integer. Not supported. Use a set of IP permissions to specify the port.[optional]}

\item{IpProtocol}{Character. Not supported. Use a set of IP permissions to specify the protocol name or number.[optional]}

\item{ToPort}{Integer. Not supported. Use a set of IP permissions to specify the port.[optional]}

\item{SourceSecurityGroupName}{Character. Not supported. Use a set of IP permissions to specify a destination security group.[optional]}

\item{SourceSecurityGroupOwnerId}{Character. Not supported. Use a set of IP permissions to specify a destination security group.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Revoke Security Group Egress
}
\section{GroupId}{

The ID of
the security group.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{IpPermissions}{

The sets of IP permissions.
You can\'t specify a destination security group and a CIDR IP address
range in the same set of permissions.
}

\section{CidrIp}{

Not supported.
Use a set of IP permissions to specify the CIDR.
}

\section{FromPort}{

Not
supported. Use a set of IP permissions to specify the
port.
}

\section{IpProtocol}{

Not supported. Use a set of IP permissions to
specify the protocol name or number.
}

\section{ToPort}{

Not supported.
Use a set of IP permissions to specify the port.
}

\section{SourceSecurityGroupName}{

Not
supported. Use a set of IP permissions to specify a destination security
group.
}

\section{SourceSecurityGroupOwnerId}{

Not supported. Use a set of IP permissions to
specify a destination security group.
}

