% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_send_diagnostic_interrupt}
\alias{ec2_send_diagnostic_interrupt}
\title{Send Diagnostic Interrupt}
\usage{
ec2_send_diagnostic_interrupt(
  InstanceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{String. The ID of the instance.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Sends a diagnostic interrupt to the specified Amazon EC2 instance to
trigger a \emph{kernel panic} (on Linux instances), or a \emph{blue screen}/\emph{stop
error} (on Windows instances). For instances based on Intel and AMD
processors, the interrupt is received as a \emph{non-maskable interrupt}
(NMI).
}
\section{InstanceId}{

The ID of the instance.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

