% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_start_instances}
\alias{ec2_start_instances}
\title{Start Instances}
\usage{
ec2_start_instances(
  InstanceId,
  AdditionalInfo = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{List or named list. The IDs of the instances.}

\item{AdditionalInfo}{String. Reserved. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Starts an Amazon EBS-backed instance that you\'ve previously stopped.
}
\section{InstanceId}{

The IDs of the instances.
}

\section{AdditionalInfo}{

Reserved.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

