% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_start_network_insights_analysis}
\alias{ec2_start_network_insights_analysis}
\title{Start Network Insights Analysis}
\usage{
ec2_start_network_insights_analysis(
  NetworkInsightsPathId,
  ClientToken,
  FilterInArn = NULL,
  DryRun = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{NetworkInsightsPathId}{String. The ID of the path.}

\item{ClientToken}{String. Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{FilterInArn}{List or named list. The Amazon Resource Names (ARN) of the resources that the path must traverse. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{TagSpecification}{List or named list. The tags to apply. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Starts analyzing the specified path. If the path is reachable, the
operation returns the shortest feasible path.
}
\section{NetworkInsightsPathId}{

The ID of the path.
}

\section{ClientToken}{

Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{How to Ensure Idempotency}.
}

\section{FilterInArn}{

The Amazon Resource Names (ARN) of the resources that the path must
traverse.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{TagSpecification}{

The tags to apply.
}

