% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_start_task}
\alias{ec2_start_task}
\title{Start Task}
\usage{
ec2_start_task(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Start Task
}
