% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_submit_task_state_change}
\alias{ec2_submit_task_state_change}
\title{Submit Task State Change}
\usage{
ec2_submit_task_state_change(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Submit Task State Change
}
