% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ec2_tag_resource}
\alias{ec2_tag_resource}
\title{Tag Resource}
\usage{
ec2_tag_resource(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Associates the specified tags to a resource with
the specified \code{resourceArn}. If existing tags on a resource are not
specified in the request parameters, they are not changed. When a
resource is deleted, the tags associated with that resource are deleted
as well.
}
