% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_terminate_client_vpn_connections}
\alias{ec2_terminate_client_vpn_connections}
\title{Terminate Client Vpn Connections}
\usage{
ec2_terminate_client_vpn_connections(
  ClientVpnEndpointId,
  ConnectionId = NULL,
  Username = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{String. The ID of the Client VPN endpoint to which the client is connected.}

\item{ConnectionId}{String. The ID of the client connection to be terminated. (optional)}

\item{Username}{String. The name of the user who initiated the connection.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Terminates active Client VPN endpoint connections. This action can be
used to terminate a specific client connection, or up to five
connections established by a specific user.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN endpoint to which the client is connected.
}

\section{ConnectionId}{

The ID of the client connection to be terminated.
}

\section{Username}{

The name of the user who initiated the connection. Use this option to
terminate all active connections for the specified user. This option can
only be used if the user has established up to five connections.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

