% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_terminate_instances}
\alias{ec2_terminate_instances}
\title{Terminate Instances}
\usage{
ec2_terminate_instances(
  InstanceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{List or named list. The IDs of the instances.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Shuts down the specified instances. This operation is idempotent; if you
terminate an instance more than once, each call succeeds.
}
\section{InstanceId}{

The IDs of the instances.

Constraints: Up to 1000 instance IDs. We recommend breaking up this
request into smaller batches.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

