% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_update_security_group_rule_descriptions_ingress}
\alias{ec2_update_security_group_rule_descriptions_ingress}
\title{Update Security Group Rule Descriptions Ingress}
\usage{
ec2_update_security_group_rule_descriptions_ingress(
  IpPermissions,
  DryRun = NULL,
  GroupId = NULL,
  GroupName = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{IpPermissions}{List or named list. The IP permissions for the security group rule.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{GroupId}{String. The ID of the security group.(optional)}

\item{GroupName}{String. [EC2-Classic, default VPC] The name of the security group.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Updates the description of an ingress (inbound) security group rule. You
can replace an existing description, or add a description to a rule that
did not have one previously.
}
\section{IpPermissions}{

The IP permissions for the security group rule.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{GroupId}{

The ID of the security group. You must specify either the security group
ID or the security group name in the request. For security groups in a
nondefault VPC, you must specify the security group ID.
}

\section{GroupName}{

[EC2-Classic, default VPC] The name of the security group. You must
specify either the security group ID or the security group name in the
request.
}

