% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_delete_account_setting}
\alias{ecs_delete_account_setting}
\title{Delete Account Setting}
\usage{
ecs_delete_account_setting(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Disables an account setting for a specified
IAM user, IAM role, or the root user for an
account.
}
