% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_delete_cluster}
\alias{ecs_delete_cluster}
\title{Delete Cluster}
\usage{
ecs_delete_cluster(cluster = NULL, simplify = TRUE, others = list())
}
\arguments{
\item{cluster}{String. The short name or full Amazon Resource Name (ARN) of the cluster to delete.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Delete Cluster
}
\section{cluster}{

The short name or full Amazon Resource Name
(ARN) of the cluster to delete.
}

