% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_describe_task_definition}
\alias{ecs_describe_task_definition}
\title{Describe Task Definition}
\usage{
ecs_describe_task_definition(
  taskDefinition = NULL,
  include = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{taskDefinition}{Character. The \code{family} for the latest \code{ACTIVE} revision, \code{family} and \code{revision} (\code{family:revision}) for a...}

\item{include}{List. Specifies whether to see the resource tags for the task definition.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Describe Task Definition
}
\section{taskDefinition}{

The \code{family} for the
latest \code{ACTIVE} revision, \code{family} and \code{revision} (\code{family:revision})
for a specific revision in the family, or full Amazon Resource Name
(ARN) of the task definition to describe.
}

\section{include}{

Specifies
whether to see the resource tags for the task definition. If \code{TAGS} is
specified, the tags are included in the response. If this field is
omitted, tags are not included in the response.
}

