% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_list_attributes}
\alias{ecs_list_attributes}
\title{List Attributes}
\usage{
ecs_list_attributes(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Lists the attributes for Amazon ECS
resources within a specified target type and cluster. When you specify a
target type and cluster, \code{ListAttributes} returns a list of attribute
objects, one for each attribute on each resource. You can filter the
list of results to a single attribute name to only return results that
have that name. You can also filter the results by attribute name and
value, for example, to see which container instances in a cluster are
running a Linux AMI (\code{ecs.os-type=linux}).
}
\section{maxResults}{

Pagination
limit
}

