% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_list_tags_for_resource}
\alias{ecs_list_tags_for_resource}
\title{List Tags For Resource}
\usage{
ecs_list_tags_for_resource(
  resourceArn = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{resourceArn}{Character. The Amazon Resource Name (ARN) that identifies the resource for which to list the tags.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
List the tags for an Amazon ECS
resource.
}
\section{resourceArn}{

The Amazon Resource Name (ARN) that identifies
the resource for which to list the tags. Currently, the supported
resources are Amazon ECS tasks, services, task definitions, clusters,
and container instances.
}

