% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_put_account_setting_default}
\alias{ecs_put_account_setting_default}
\title{Put Account Setting Default}
\usage{
ecs_put_account_setting_default(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Modifies an account setting for all IAM users on an
account for whom no individual account setting has been specified.
Account settings are set on a per-Region
basis.
}
