% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_put_attributes}
\alias{ecs_put_attributes}
\title{Put Attributes}
\usage{
ecs_put_attributes(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Create or update an attribute on an Amazon
ECS resource. If the attribute does not exist, it is created. If the
attribute exists, its value is replaced with the specified value. To
delete an attribute, use DeleteAttributes. For more information, see
\href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-placement-constraints.html#attributes}{Attributes}
in the \emph{Amazon Elastic Container Service Developer
Guide}.
}
