% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_submit_attachment_state_changes}
\alias{ecs_submit_attachment_state_changes}
\title{Submit Attachment State Changes}
\usage{
ecs_submit_attachment_state_changes(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Submit Attachment State Changes
}
