% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_array}
\alias{list_to_array}
\title{Utility functions}
\usage{
list_to_array(prefix, x)
}
\arguments{
\item{prefix}{Character, the name of the parameter}

\item{x}{List or Vector, the value of the parameter}
}
\value{
A list object
}
\description{
Utility functions. \code{list_to_array} converts a list or a vector to an
Array object in AWS Documentation. \code{list_to_filter} converts a filter
list to an Array object.
}
\examples{
vpc_id = c("vpc-081ec835f3", "vpc-0ee975135d", "vpc-06e4ab6c6c")
list_to_array("VpcId", vpc_id)

filter = list(`dhcp-options-id` = c("dopt-7a8b9c2d", "dopt-2b2a3d3c"), state="available")
list_to_filter(filter)
}
