% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-settings.R
\name{aws_set_retry_time}
\alias{aws_set_retry_time}
\alias{aws_set_print_on_error}
\alias{aws_set_network_timeout}
\alias{aws_get_retry_time}
\alias{aws_get_print_on_error}
\alias{aws_get_network_timeout}
\title{Get or set the package settings}
\usage{
aws_set_retry_time(x)

aws_set_print_on_error(x)

aws_set_network_timeout(x)

aws_get_retry_time()

aws_get_print_on_error()

aws_get_network_timeout()
}
\arguments{
\item{x}{the value to be set. For the timeout setting, the unit is seconds.}
}
\value{
Setter: The old value
Getter: The current value
}
\description{
Get or set the package settings
}
\examples{
## Set the timeout to 10 seconds
aws_set_network_timeout(10)
## Get the timeout setting
aws_get_network_timeout()

## Turn off print on error
aws_set_print_on_error(FALSE)

## Set the retry times to 5
aws_set_retry_time(5)

}
