% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions-documentation.R, R/ec2-actions.R,
%   R/ecs-actions.R
\name{ecs_run_task}
\alias{ecs_run_task}
\alias{ec2_create_internet_gateway}
\alias{ec2_delete_internet_gateway}
\alias{ec2_attach_internet_gateway}
\alias{ec2_detach_internet_gateway}
\alias{ec2_create_security_group}
\alias{ec2_delete_security_group}
\alias{ec2_authorize_security_group_ingress}
\alias{ec2_create_vpc}
\alias{ec2_delete_vpc}
\alias{ec2_create_subnet}
\alias{ec2_delete_subnet}
\alias{ec2_create_route}
\alias{ec2_delete_route}
\alias{ec2_create_route_table}
\alias{ec2_delete_route_table}
\alias{ec2_describe_network_interfaces}
\alias{ec2_describe_internet_gateways}
\alias{ec2_describe_route_tables}
\alias{ec2_describe_security_groups}
\alias{ec2_describe_vpcs}
\alias{ec2_describe_subnets}
\alias{ecs_register_task_definition}
\alias{ecs_deregister_task_definition}
\alias{ecs_create_cluster}
\alias{ecs_delete_cluster}
\alias{ecs_stop_task}
\alias{ecs_describe_tasks}
\alias{ecs_list_task_definitions}
\alias{ecs_list_clusters}
\alias{ecs_list_tasks}
\title{AWS actions}
\usage{
ec2_create_internet_gateway(json = list())

ec2_delete_internet_gateway(json = list())

ec2_attach_internet_gateway(json = list())

ec2_detach_internet_gateway(json = list())

ec2_create_security_group(json = list())

ec2_delete_security_group(json = list())

ec2_authorize_security_group_ingress(json = list())

ec2_create_vpc(json = list())

ec2_delete_vpc(json = list())

ec2_create_subnet(json = list())

ec2_delete_subnet(json = list())

ec2_create_route(json = list())

ec2_delete_route(json = list())

ec2_create_route_table(json = list())

ec2_delete_route_table(json = list())

ec2_describe_network_interfaces(json = list())

ec2_describe_internet_gateways(json = list())

ec2_describe_route_tables(json = list())

ec2_describe_security_groups(json = list())

ec2_describe_vpcs(json = list())

ec2_describe_subnets(json = list())

ecs_register_task_definition(json = list())

ecs_deregister_task_definition(json = list())

ecs_create_cluster(json = list())

ecs_delete_cluster(json = list())

ecs_run_task(json = list())

ecs_stop_task(json = list())

ecs_describe_tasks(json = list())

ecs_list_task_definitions(json = list())

ecs_list_clusters(json = list())

ecs_list_tasks(json = list())
}
\arguments{
\item{json}{A list object containing the request parameters}
}
\value{
A vector/list object depending on the API type
}
\description{
AWS actions for communicating with ECS.
}
\details{
These functions provide a minimum set of APIs for controlling the ECS instance.
Please refer to \href{https://docs.aws.amazon.com/index.html}{AWS Documentation}
for the API documentation. The credentials must be provided before sending the
REST request. The credentials can be found by the environment variables or
set via \code{aws_set_credentials}.

For the actions which use \code{nextToken} to return the result by parts
(e.g. \code{ecs_list_task_definitions}), the token will be handled
automatically and the complete result will be returned by the function. There is
no need to include \code{nextToken} parameter in your request.

If you find any ECS API useful but not listed here, please consider to submit an issue
to the package repository described in the DESCRIPTION file.
}
