% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions-documentation.R
\name{ecs_run_task}
\alias{ecs_run_task}
\title{AWS actions}
\value{
A vector/list object depending on the API type
}
\description{
AWS actions for communicating with ECS.
}
\details{
These functions provide a set of APIs for controlling the ECS instance.
Please refer to \href{https://docs.aws.amazon.com/index.html}{AWS Documentation}
for the API documentation. The credentials must be provided before sending the
REST request. The credentials can be found by the environment variables or
set via \code{aws_set_credentials}.

For the actions which use \code{nextToken} to return the result by parts
(e.g. \code{ecs_list_task_definitions}), the token can be handled
automatically and the complete result can be returned by the function
if the argument \code{simplify = TRUE}.

If you find any ECS API useful but not listed here, please consider to submit an issue
to the package repository described in the DESCRIPTION file.
}
