% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_address}
\alias{ec2_associate_address}
\title{Associate Address}
\usage{
ec2_associate_address(
  AllocationId = NULL,
  InstanceId = NULL,
  PublicIp = NULL,
  AllowReassociation = NULL,
  DryRun = NULL,
  NetworkInterfaceId = NULL,
  PrivateIpAddress = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{AllocationId}{Character. [EC2-VPC] The allocation ID. This is required for EC2-VPC.[optional]}

\item{InstanceId}{Character. The ID of the instance.[optional]}

\item{PublicIp}{Character. [EC2-Classic] The Elastic IP address to associate with the instance.[optional]}

\item{AllowReassociation}{Logical. [EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that...[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{NetworkInterfaceId}{Character.   [EC2-VPC] The ID of the network interface.[optional]}

\item{PrivateIpAddress}{Character. [EC2-VPC] The primary or secondary private IP address to associate with the Elastic IP address.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Associate Address
}
\section{AllocationId}{

[EC2-VPC] The
allocation ID. This is required for EC2-VPC.
}

\section{InstanceId}{

The ID
of the instance. The instance must have exactly one attached network
interface. For EC2-VPC, you can specify either the instance ID or the
network interface ID, but not both. For EC2-Classic, you must specify an
instance ID and the instance must be in the running
state.
}

\section{PublicIp}{

[EC2-Classic] The Elastic IP address to
associate with the instance. This is required for
EC2-Classic.
}

\section{AllowReassociation}{

[EC2-VPC] For a VPC in an EC2-Classic
account, specify true to allow an Elastic IP address that is already
associated with an instance or network interface to be reassociated with
the specified instance or network interface. Otherwise, the operation
fails. In a VPC in an EC2-VPC-only account, reassociation is automatic,
therefore you can specify false to ensure the operation fails if the
Elastic IP address is already associated with another
resource.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{NetworkInterfaceId}{


[EC2-VPC] The ID of the network interface. If the instance has more
than one network interface, you must specify a network interface ID.

For EC2-VPC, you can specify either the instance ID or the network
interface ID, but not both.
}

\section{PrivateIpAddress}{

[EC2-VPC] The primary or secondary private IP
address to associate with the Elastic IP address. If no private IP
address is specified, the Elastic IP address is associated with the
primary private IP address.
}

