% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_dhcp_options}
\alias{ec2_associate_dhcp_options}
\title{Associate Dhcp Options}
\usage{
ec2_associate_dhcp_options(
  DhcpOptionsId,
  VpcId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DhcpOptionsId}{String. The ID of the DHCP options set, or \code{default} to associate no DHCP options with the VPC.}

\item{VpcId}{String. The ID of the VPC.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Associate Dhcp Options
}
\section{DhcpOptionsId}{

The ID of
the DHCP options set, or \code{default} to associate no DHCP options with the
VPC.
}

\section{VpcId}{

The ID of the VPC.
}

\section{DryRun}{

Checks
whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

