% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_iam_instance_profile}
\alias{ec2_associate_iam_instance_profile}
\title{Associate Iam Instance Profile}
\usage{
ec2_associate_iam_instance_profile(
  IamInstanceProfile,
  InstanceId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{IamInstanceProfile}{Object. The IAM instance profile.}

\item{InstanceId}{Character. The ID of the instance.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Associates an IAM instance profile with a running
or stopped instance. You cannot associate more than one IAM instance
profile with an instance.
}
\section{IamInstanceProfile}{

The IAM instance
profile.
}

\section{InstanceId}{

The ID of the
instance.
}

