% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_route_table}
\alias{ec2_associate_route_table}
\title{Associate Route Table}
\usage{
ec2_associate_route_table(
  RouteTableId,
  DryRun = NULL,
  SubnetId = NULL,
  GatewayId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{RouteTableId}{String. The ID of the route table.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{SubnetId}{String. The ID of the subnet. (optional)}

\item{GatewayId}{String. The ID of the internet gateway or virtual private gateway. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Associates a subnet in your VPC or an internet gateway or virtual
private gateway attached to your VPC with a route table in your VPC.
This association causes traffic from the subnet or gateway to be routed
according to the routes in the route table. The action returns an
association ID, which you need in order to disassociate the route table
later. A route table can be associated with multiple subnets.
}
\section{RouteTableId}{

The ID of the route table.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{SubnetId}{

The ID of the subnet.
}

\section{GatewayId}{

The ID of the internet gateway or virtual private gateway.
}

