% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_transit_gateway_multicast_domain}
\alias{ec2_associate_transit_gateway_multicast_domain}
\title{Associate Transit Gateway Multicast Domain}
\usage{
ec2_associate_transit_gateway_multicast_domain(
  TransitGatewayMulticastDomainId = NULL,
  TransitGatewayAttachmentId = NULL,
  SubnetIds = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayMulticastDomainId}{String. The ID of the transit gateway multicast domain. (optional)}

\item{TransitGatewayAttachmentId}{String. The ID of the transit gateway attachment to associate with the transit gateway multicast domain. (optional)}

\item{SubnetIds}{List or named list. The IDs of the subnets to associate with the transit gateway multicast domain. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Associates the specified subnets and transit gateway attachments with
the specified transit gateway multicast domain.
}
\section{TransitGatewayMulticastDomainId}{

The ID of the transit gateway multicast domain.
}

\section{TransitGatewayAttachmentId}{

The ID of the transit gateway attachment to associate with the transit
gateway multicast domain.
}

\section{SubnetIds}{

The IDs of the subnets to associate with the transit gateway multicast
domain.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

