% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_associate_transit_gateway_route_table}
\alias{ec2_associate_transit_gateway_route_table}
\title{Associate Transit Gateway Route Table}
\usage{
ec2_associate_transit_gateway_route_table(
  TransitGatewayRouteTableId,
  TransitGatewayAttachmentId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayRouteTableId}{String. The ID of the transit gateway route table.}

\item{TransitGatewayAttachmentId}{String. The ID of the attachment.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Associates the specified attachment with the
specified transit gateway route table. You can associate only one route
table with an attachment.
}
\section{TransitGatewayRouteTableId}{

The ID of the transit
gateway route table.
}

\section{TransitGatewayAttachmentId}{

The ID of the
attachment.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

