% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_attach_classic_link_vpc}
\alias{ec2_attach_classic_link_vpc}
\title{Attach Classic Link Vpc}
\usage{
ec2_attach_classic_link_vpc(
  SecurityGroupId,
  InstanceId,
  VpcId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{SecurityGroupId}{List or named list. The ID of one or more of the VPC\'s security groups.}

\item{InstanceId}{String. The ID of an EC2-Classic instance to link to the ClassicLink-enabled VPC.}

\item{VpcId}{String. The ID of a ClassicLink-enabled VPC.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Links an EC2-Classic instance to a ClassicLink-enabled VPC through one
or more of the VPC\'s security groups. You cannot link an EC2-Classic
instance to more than one VPC at a time. You can only link an instance
that\'s in the \code{running} state. An instance is automatically unlinked
from a VPC when it\'s stopped - you can link it to the VPC again when
you restart it.
}
\section{SecurityGroupId}{

The ID of one or more of the VPC\'s security groups. You cannot specify
security groups from a different VPC.
}

\section{InstanceId}{

The ID of an EC2-Classic instance to link to the ClassicLink-enabled
VPC.
}

\section{VpcId}{

The ID of a ClassicLink-enabled VPC.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

