% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_bundle_instance}
\alias{ec2_bundle_instance}
\title{Bundle Instance}
\usage{
ec2_bundle_instance(
  InstanceId,
  Storage,
  DryRun = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout(),
  region = aws_get_region()
)
}
\arguments{
\item{InstanceId}{Character.   The ID of the instance to bundle.  Type: String  Default: None  Required: Yes}

\item{Storage}{Object. The bucket in which to store the AMI.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}

\item{region}{Character. The region of the AWS service.}
}
\value{
A list object or a character vector
}
\description{
Bundle Instance
}
\section{InstanceId}{


The ID of the instance to bundle.

Type: String

Default: None

Required: Yes
}

\section{Storage}{

The bucket in which to store the AMI. You can
specify a bucket that you already own or a new bucket that Amazon EC2
creates on your behalf. If you specify a bucket that belongs to someone
else, Amazon EC2 returns an error.
}

\section{DryRun}{

Checks whether
you have the required permissions for the action, without actually
making the request, and provides an error response. If you have the
required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

