% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_capacity_reservation}
\alias{ec2_cancel_capacity_reservation}
\title{Cancel Capacity Reservation}
\usage{
ec2_cancel_capacity_reservation(
  CapacityReservationId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{CapacityReservationId}{Character. The ID of the Capacity Reservation to be cancelled.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Cancel Capacity Reservation
}
\section{CapacityReservationId}{

The ID of the Capacity
Reservation to be cancelled.
}

\section{DryRun}{

Checks whether you have
the required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

