% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_export_task}
\alias{ec2_cancel_export_task}
\title{Cancel Export Task}
\usage{
ec2_cancel_export_task(
  ExportTaskId,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{ExportTaskId}{Character. The ID of the export task. This is the ID returned by \code{CreateInstanceExportTask}.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Cancels an active export task. The request removes
all artifacts of the export, including any partially-created Amazon S3
objects. If the export task is complete or is in the process of
transferring the final disk image, the command fails and returns an
error.
}
\section{ExportTaskId}{

The ID of the export task. This
is the ID returned by
\code{CreateInstanceExportTask}.
}

