% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_cancel_reserved_instances_listing}
\alias{ec2_cancel_reserved_instances_listing}
\title{Cancel Reserved Instances Listing}
\usage{
ec2_cancel_reserved_instances_listing(
  ReservedInstancesListingId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ReservedInstancesListingId}{String. The ID of the Reserved Instance listing.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Cancels the specified Reserved Instance listing in the Reserved Instance
Marketplace.
}
\section{ReservedInstancesListingId}{

The ID of the Reserved Instance listing.
}

