% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_local_gateway_route_table_vpc_association}
\alias{ec2_create_local_gateway_route_table_vpc_association}
\title{Create Local Gateway Route Table Vpc Association}
\usage{
ec2_create_local_gateway_route_table_vpc_association(
  LocalGatewayRouteTableId,
  VpcId,
  TagSpecification = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{LocalGatewayRouteTableId}{Character. The ID of the local gateway route table.}

\item{VpcId}{Character. The ID of the VPC.}

\item{TagSpecification}{List. The tags to assign to the local gateway route table VPC association.[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Associates the specified VPC
with the specified local gateway route table.
}
\section{LocalGatewayRouteTableId}{

The ID of the local gateway
route table.
}

\section{VpcId}{

The ID of the
VPC.
}

\section{TagSpecification}{

The tags to assign to the local gateway route
table VPC association.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

