% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway}
\alias{ec2_create_transit_gateway}
\title{Create Transit Gateway}
\usage{
ec2_create_transit_gateway(
  Description = NULL,
  Options = NULL,
  TagSpecification = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Description}{String. A description of the transit gateway. (optional)}

\item{Options}{Object. The transit gateway options. (optional)}

\item{TagSpecification}{List or named list. The tags to apply to the transit gateway. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates a transit gateway.
}
\section{Description}{

A description of the transit gateway.
}

\section{Options}{

The transit gateway options.
}

\section{TagSpecification}{

The tags to apply to the transit gateway.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

