% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_transit_gateway_route_table}
\alias{ec2_create_transit_gateway_route_table}
\title{Create Transit Gateway Route Table}
\usage{
ec2_create_transit_gateway_route_table(
  TransitGatewayId,
  TagSpecifications = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayId}{String. The ID of the transit gateway.}

\item{TagSpecifications}{List. The tags to apply to the transit gateway route table.[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Creates a route table for the specified
transit gateway.
}
\section{TransitGatewayId}{

The ID of the transit
gateway.
}

\section{TagSpecifications}{

The tags to apply to the transit gateway
route table.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

