% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_create_vpn_connection_route}
\alias{ec2_create_vpn_connection_route}
\title{Create Vpn Connection Route}
\usage{
ec2_create_vpn_connection_route(
  DestinationCidrBlock,
  VpnConnectionId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DestinationCidrBlock}{String. The CIDR block associated with the local subnet of the customer network.}

\item{VpnConnectionId}{String. The ID of the VPN connection.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Create Vpn Connection Route
}
\section{DestinationCidrBlock}{

The CIDR block associated with the local
subnet of the customer network.
}

\section{VpnConnectionId}{

The ID of the VPN
connection.
}

