% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_fleets}
\alias{ec2_delete_fleets}
\title{Delete Fleets}
\usage{
ec2_delete_fleets(
  FleetId,
  TerminateInstances,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FleetId}{List or named list. The IDs of the EC2 Fleets.}

\item{TerminateInstances}{Boolean. Indicates whether to terminate the instances when the EC2 Fleet is deleted.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified EC2 Fleet.
}
\section{FleetId}{

The IDs of the EC2 Fleets.
}

\section{TerminateInstances}{

Indicates whether to terminate the instances when the EC2 Fleet is
deleted. The default is to terminate the instances.

To let the instances continue to run after the EC2 Fleet is deleted,
specify \code{NoTerminateInstances}. Supported only for fleets of type
\code{maintain} and \code{request}.

For \code{instant} fleets, you cannot specify \code{NoTerminateInstances}. A
deleted \code{instant} fleet with running instances is not supported.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

