% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_flow_logs}
\alias{ec2_delete_flow_logs}
\title{Delete Flow Logs}
\usage{
ec2_delete_flow_logs(
  FlowLogId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FlowLogId}{List.   One or more flow log IDs.  Constraint: Maximum of 1000 flow log IDs.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Deletes one or more flow
logs.
}
\section{FlowLogId}{


One or more flow log IDs.

Constraint: Maximum of 1000 flow log IDs.
}

\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

