% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_local_gateway_route}
\alias{ec2_delete_local_gateway_route}
\title{Delete Local Gateway Route}
\usage{
ec2_delete_local_gateway_route(
  DestinationCidrBlock,
  LocalGatewayRouteTableId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DestinationCidrBlock}{String. The CIDR range for the route. This must match the CIDR for the route exactly.}

\item{LocalGatewayRouteTableId}{String. The ID of the local gateway route table.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified route
from the specified local gateway route
table.
}
\section{DestinationCidrBlock}{

The CIDR range for the
route. This must match the CIDR for the route
exactly.
}

\section{LocalGatewayRouteTableId}{

The ID of the local gateway route
table.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

