% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_network_acl}
\alias{ec2_delete_network_acl}
\title{Delete Network Acl}
\usage{
ec2_delete_network_acl(
  NetworkAclId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{NetworkAclId}{String. The ID of the network ACL.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified network ACL. You
can\'t delete the ACL if it\'s associated with any subnets. You can\'t
delete the default network ACL.
}
\section{NetworkAclId}{

The ID of the network
ACL.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

