% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_queued_reserved_instances}
\alias{ec2_delete_queued_reserved_instances}
\title{Delete Queued Reserved Instances}
\usage{
ec2_delete_queued_reserved_instances(
  ReservedInstancesId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ReservedInstancesId}{List or named list. The IDs of the Reserved Instances.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the queued purchases for the specified Reserved Instances.
}
\section{ReservedInstancesId}{

The IDs of the Reserved Instances.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

