% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_security_group}
\alias{ec2_delete_security_group}
\title{Delete Security Group}
\usage{
ec2_delete_security_group(
  GroupId = NULL,
  GroupName = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{GroupId}{String. The ID of the security group. Required for a nondefault VPC. (optional)}

\item{GroupName}{String. [EC2-Classic, default VPC] The name of the security group.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes a security group.
}
\section{GroupId}{

The ID of the security group. Required for a nondefault VPC.
}

\section{GroupName}{

[EC2-Classic, default VPC] The name of the security group. You can
specify either the security group name or the security group ID.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

