% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_delete_vpn_connection_route}
\alias{ec2_delete_vpn_connection_route}
\title{Delete Vpn Connection Route}
\usage{
ec2_delete_vpn_connection_route(
  DestinationCidrBlock,
  VpnConnectionId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DestinationCidrBlock}{String. The CIDR block associated with the local subnet of the customer network.}

\item{VpnConnectionId}{String. The ID of the VPN connection.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deletes the specified static route associated with a VPN connection
between an existing virtual private gateway and a VPN customer gateway.
The static route allows traffic to be routed from the virtual private
gateway to the VPN customer gateway.
}
\section{DestinationCidrBlock}{

The CIDR block associated with the local subnet of the customer network.
}

\section{VpnConnectionId}{

The ID of the VPN connection.
}

