% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_deregister_image}
\alias{ec2_deregister_image}
\title{Deregister Image}
\usage{
ec2_deregister_image(ImageId, DryRun = NULL, simplify = TRUE, others = list())
}
\arguments{
\item{ImageId}{String. The ID of the AMI.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Deregisters the specified AMI. After you deregister an AMI, it can\'t be
used to launch new instances; however, it doesn\'t affect any instances
that you\'ve already launched from the AMI. You\'ll continue to incur
usage costs for those instances until you terminate them.
}
\section{ImageId}{

The ID of the AMI.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

