% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_deregister_transit_gateway_multicast_group_sources}
\alias{ec2_deregister_transit_gateway_multicast_group_sources}
\title{Deregister Transit Gateway Multicast Group Sources}
\usage{
ec2_deregister_transit_gateway_multicast_group_sources(
  TransitGatewayMulticastDomainId = NULL,
  GroupIpAddress = NULL,
  NetworkInterfaceIds = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayMulticastDomainId}{String. The ID of the transit gateway multicast domain.\link{optional}}

\item{GroupIpAddress}{String. The IP address assigned to the transit gateway multicast group.\link{optional}}

\item{NetworkInterfaceIds}{Array. The IDs of the group sources\' network interfaces.\link{optional}}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Deregisters the specified sources
(network interfaces) from the transit gateway multicast
group.
}
\section{TransitGatewayMulticastDomainId}{

The ID of
the transit gateway multicast domain.
}

\section{GroupIpAddress}{

The IP address
assigned to the transit gateway multicast group.
}

\section{NetworkInterfaceIds}{

The
IDs of the group sources\' network
interfaces.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

