% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_client_vpn_authorization_rules}
\alias{ec2_describe_client_vpn_authorization_rules}
\title{Describe Client Vpn Authorization Rules}
\usage{
ec2_describe_client_vpn_authorization_rules(
  ClientVpnEndpointId,
  DryRun = NULL,
  NextToken = NULL,
  Filter = NULL,
  MaxResults = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{String. The ID of the Client VPN endpoint.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the authorization rules for a specified Client VPN endpoint.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN endpoint.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Filter}{

One or more filters. Filter names and values are case-sensitive.
\itemize{
\item \code{description} - The description of the authorization rule.
\item \code{destination-cidr} - The CIDR of the network to which the
authorization rule applies.
\item \code{group-id} - The ID of the Active Directory group to which the
authorization rule grants access.
}
}

\section{MaxResults}{

The maximum number of results to return for the request in a single
page. The remaining results can be seen by sending another request with
the nextToken value.
}

