% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_coip_pools}
\alias{ec2_describe_coip_pools}
\title{Describe Coip Pools}
\usage{
ec2_describe_coip_pools(
  PoolId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{PoolId}{Array. The IDs of the address pools.\link{optional}}

\item{Filter}{No description can be found.}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.\link{optional}}

\item{NextToken}{Characters, The token for the next page of results.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified customer-owned
address pools or all of your customer-owned address
pools.
}
\section{PoolId}{

The IDs of the address
pools.
}

\section{Filter}{


The filters. The following are the possible values:
\itemize{
\item \code{coip-pool.pool-id}
}\if{html}{\out{<div class="{=html}">}}\preformatted{
}\if{html}{\out{</div>}}
\itemize{
\item \code{coip-pool.local-gateway-route-table-id}
}
}

\section{MaxResults}{

The maximum number of results to return with a
single call. To retrieve the remaining results, make another call with
the returned \code{nextToken} value.
}

\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

