% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_dhcp_options}
\alias{ec2_describe_dhcp_options}
\title{Describe Dhcp Options}
\usage{
ec2_describe_dhcp_options(
  DhcpOptionsId = NULL,
  Filter = NULL,
  DryRun = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DhcpOptionsId}{List.   The IDs of one or more DHCP options sets.  Default: Describes all your DHCP options sets.  [optional]}

\item{Filter}{Named list where the name is the filter name and the value is the
filter's value. The value can be a vector or a list object. (see below)[optional]}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{NextToken}{Characters, The token for the next page of results.}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describe Dhcp Options
}
\section{DhcpOptionsId}{


The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.
}

\section{Filter}{


One or more filters.
\itemize{
\item \code{dhcp-options-id} - The ID of a DHCP options set.
\item \code{key} - The key for one of the options (for example, \code{domain-name}).
\item \code{value} - The value for one of the options.
\item \code{owner-id} - The ID of the AWS account that owns the DHCP options
set.
\item \code{tag}:\ - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}
}

\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{MaxResults}{

The maximum number of results to return
with a single call. To retrieve the remaining results, make another call
with the returned \code{nextToken} value.
}

