% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_elastic_gpus}
\alias{ec2_describe_elastic_gpus}
\title{Describe Elastic Gpus}
\usage{
ec2_describe_elastic_gpus(
  ElasticGpuId = NULL,
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ElasticGpuId}{Array. The Elastic Graphics accelerator IDs.\link{optional}}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{Filter}{No description can be found.}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.\link{optional}}

\item{NextToken}{Characters, The token for the next page of results.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes the Elastic Graphics accelerator
associated with your instances. For more information about Elastic
Graphics, see \href{https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/elastic-graphics.html}{Amazon Elastic Graphics}.
}
\section{ElasticGpuId}{

The Elastic Graphics accelerator
IDs.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Filter}{


The filters.
\itemize{
\item \code{availability-zone} - The Availability Zone in which the Elastic
Graphics accelerator resides.
\item \code{elastic-gpu-health} - The status of the Elastic Graphics
accelerator (\code{OK} \| \code{IMPAIRED}).
\item \code{elastic-gpu-state} - The state of the Elastic Graphics accelerator
(\code{ATTACHED}).
\item \code{elastic-gpu-type} - The type of Elastic Graphics accelerator; for
example, \code{eg1.medium}.
\item \code{instance-id} - The ID of the instance to which the Elastic Graphics
accelerator is associated.
}
}

\section{MaxResults}{

The maximum number of results to return in a single
call. To retrieve the remaining results, make another call with the
returned \code{NextToken} value. This value can be between 5 and
1000.
}

