% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_export_tasks}
\alias{ec2_describe_export_tasks}
\title{Describe Export Tasks}
\usage{
ec2_describe_export_tasks(
  ExportTaskId = NULL,
  Filter = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ExportTaskId}{Array. The export task IDs.\link{optional}}

\item{Filter}{No description can be found.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified export instance tasks
or all of your export instance tasks.
}
\section{ExportTaskId}{

The export task
IDs.
}

\section{Filter}{

the filters for the export
tasks.
}

