% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_fpga_image_attribute}
\alias{ec2_describe_fpga_image_attribute}
\title{Describe Fpga Image Attribute}
\usage{
ec2_describe_fpga_image_attribute(
  FpgaImageId,
  Attribute,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{FpgaImageId}{String. The ID of the AFI.}

\item{Attribute}{String. The AFI attribute.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Describes the specified attribute of the
specified Amazon FPGA Image (AFI).
}
\section{FpgaImageId}{

The ID of the
AFI.
}

\section{Attribute}{

The AFI attribute.
}

\section{DryRun}{

Checks
whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

