% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_iam_instance_profile_associations}
\alias{ec2_describe_iam_instance_profile_associations}
\title{Describe Iam Instance Profile Associations}
\usage{
ec2_describe_iam_instance_profile_associations(
  AssociationId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{Array. The IAM instance profile associations.\link{optional}}

\item{Filter}{No description can be found.}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.\link{optional}}

\item{NextToken}{Characters, The token for the next page of results.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes your IAM instance profile
associations.
}
\section{AssociationId}{

The IAM instance profile
associations.
}

\section{Filter}{


The filters.
\itemize{
\item \code{instance-id} - The ID of the instance.
\item \code{state} - The state of the association (\code{associating} \|
\code{associated} \| \code{disassociating}).
}
}

\section{MaxResults}{

The maximum number of results to return in a single
call. To retrieve the remaining results, make another call with the
returned \code{NextToken} value.
}

