% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_id_format}
\alias{ec2_describe_id_format}
\title{Describe Id Format}
\usage{
ec2_describe_id_format(
  Resource = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{Resource}{Character. The type of resource: \code{bundle} \| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options} \| \code{elastic-ip-allocation}...[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Describe Id Format
}
\section{Resource}{

The type of resource: \code{bundle}
\| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options} \|
\code{elastic-ip-allocation} \| \code{elastic-ip-association} \| \code{export-task} \|
\code{flow-log} \| \code{image} \| \code{import-task} \| \code{instance} \|
\code{internet-gateway} \| \code{network-acl} \| \code{network-acl-association} \|
\code{network-interface} \| \code{network-interface-attachment} \| \code{prefix-list}
\| \code{reservation} \| \code{route-table} \| \code{route-table-association} \|
\code{security-group} \| \code{snapshot} \| \code{subnet} \|
\code{subnet-cidr-block-association} \| \code{volume} \| \code{vpc} \|
\code{vpc-cidr-block-association} \| \code{vpc-endpoint} \|
\code{vpc-peering-connection} \| \code{vpn-connection} \| \code{vpn-gateway}
}

