% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_image_attribute}
\alias{ec2_describe_image_attribute}
\title{Describe Image Attribute}
\usage{
ec2_describe_image_attribute(
  Attribute,
  ImageId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The AMI attribute.}

\item{ImageId}{String. The ID of the AMI.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified attribute of the specified AMI. You can specify
only one attribute at a time.
}
\section{Attribute}{

The AMI attribute.

\strong{Note}: Depending on your account privileges, the \code{blockDeviceMapping}
attribute may return a \code{Client.AuthFailure} error. If this happens, use
DescribeImages to get information about the block device mapping for the
AMI.
}

\section{ImageId}{

The ID of the AMI.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

