% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_import_image_tasks}
\alias{ec2_describe_import_image_tasks}
\title{Describe Import Image Tasks}
\usage{
ec2_describe_import_image_tasks(
  DryRun = NULL,
  Filters = NULL,
  ImportTaskId = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{Filters}{List. Filter tasks using the \code{task-state} filter and one of the following values: \code{active}, \code{completed},...[optional]}

\item{ImportTaskId}{List. The IDs of the import image tasks.[optional]}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.[optional]}

\item{NextToken}{Characters. The token for the next page of results[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Displays details about an import virtual machine or
import snapshot tasks that are already
created.
}
\section{DryRun}{

Checks whether you have the
required permissions for the action, without actually making the
request, and provides an error response. If you have the required
permissions, the error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Filters}{

Filter tasks using the
\code{task-state} filter and one of the following values: \code{active},
\code{completed}, \code{deleting}, or \code{deleted}.
}

\section{ImportTaskId}{

The IDs of
the import image tasks.
}

\section{MaxResults}{

The maximum number of
results to return in a single call.
}

