% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_prefix_lists}
\alias{ec2_describe_prefix_lists}
\title{Describe Prefix Lists}
\usage{
ec2_describe_prefix_lists(
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  PrefixListId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{PrefixListId}{List or named list. One or more prefix list IDs. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes available AWS services in a prefix list format, which includes
the prefix list name and prefix list ID of the service and the IP
address range for the service.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Filter}{

One or more filters.
\itemize{
\item \code{prefix-list-id}: The ID of a prefix list.
\item \code{prefix-list-name}: The name of a prefix list.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{PrefixListId}{

One or more prefix list IDs.
}

