% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_regions}
\alias{ec2_describe_regions}
\title{Describe Regions}
\usage{
ec2_describe_regions(
  Filter = NULL,
  RegionName = NULL,
  DryRun = NULL,
  AllRegions = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{RegionName}{List or named list. The names of the Regions.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{AllRegions}{Boolean. Indicates whether to display all Regions, including Regions that are disabled for your account. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the Regions that are enabled for your account, or all Regions.
}
\section{Filter}{

The filters.
\itemize{
\item \code{endpoint} - The endpoint of the Region (for example,
\verb{ec2.us-east-1.amazonaws.com}).
\item \verb{opt-in-status} - The opt-in status of the Region
(\verb{opt-in-not-required} \| \verb{opted-in} \| \verb{not-opted-in}).
\item \code{region-name} - The name of the Region (for example, \code{us-east-1}).
}
}

\section{RegionName}{

The names of the Regions. You can specify any Regions, whether they are
enabled and disabled for your account.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{AllRegions}{

Indicates whether to display all Regions, including Regions that are
disabled for your account.
}

