% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_reserved_instances_listings}
\alias{ec2_describe_reserved_instances_listings}
\title{Describe Reserved Instances Listings}
\usage{
ec2_describe_reserved_instances_listings(
  Filter = NULL,
  ReservedInstancesId = NULL,
  ReservedInstancesListingId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{ReservedInstancesId}{String. One or more Reserved Instance IDs. (optional)}

\item{ReservedInstancesListingId}{String. One or more Reserved Instance listing IDs. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes your account\'s Reserved Instance listings in the Reserved
Instance Marketplace.
}
\section{Filter}{

One or more filters.
\itemize{
\item \code{reserved-instances-id} - The ID of the Reserved Instances.
\item \code{reserved-instances-listing-id} - The ID of the Reserved Instances
listing.
\item \code{status} - The status of the Reserved Instance listing (\code{pending} \|
\code{active} \| \code{cancelled} \| \code{closed}).
\item \code{status-message} - The reason for the status.
}
}

\section{ReservedInstancesId}{

One or more Reserved Instance IDs.
}

\section{ReservedInstancesListingId}{

One or more Reserved Instance listing IDs.
}

