% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_traffic_mirror_targets}
\alias{ec2_describe_traffic_mirror_targets}
\title{Describe Traffic Mirror Targets}
\usage{
ec2_describe_traffic_mirror_targets(
  TrafficMirrorTargetId = NULL,
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TrafficMirrorTargetId}{List or named list. The ID of the Traffic Mirror targets. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Information about one or more Traffic Mirror targets.
}
\section{TrafficMirrorTargetId}{

The ID of the Traffic Mirror targets.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Filter}{

One or more filters. The possible values are:
\itemize{
\item \code{description}: The Traffic Mirror target description.
\item \code{network-interface-id}: The ID of the Traffic Mirror session network
interface.
\item \code{network-load-balancer-arn}: The Amazon Resource Name (ARN) of the
Network Load Balancer that is associated with the session.
\item \code{owner-id}: The ID of the account that owns the Traffic Mirror
session.
\item \code{traffic-mirror-target-id}: The ID of the Traffic Mirror target.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

