% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_transit_gateway_multicast_domains}
\alias{ec2_describe_transit_gateway_multicast_domains}
\title{Describe Transit Gateway Multicast Domains}
\usage{
ec2_describe_transit_gateway_multicast_domains(
  TransitGatewayMulticastDomainIds = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayMulticastDomainIds}{List or named list. The ID of the transit gateway multicast domain. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes one or more transit gateway multicast domains.
}
\section{TransitGatewayMulticastDomainIds}{

The ID of the transit gateway multicast domain.
}

\section{Filter}{

One or more filters. The possible values are:
\itemize{
\item \code{state} - The state of the transit gateway multicast domain. Valid
values are \code{pending} \| \code{available} \| \code{deleting} \| \code{deleted}.
\item \code{transit-gateway-id} - The ID of the transit gateway.
\item \code{transit-gateway-multicast-domain-id} - The ID of the transit
gateway multicast domain.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

