% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_transit_gateway_vpc_attachments}
\alias{ec2_describe_transit_gateway_vpc_attachments}
\title{Describe Transit Gateway Vpc Attachments}
\usage{
ec2_describe_transit_gateway_vpc_attachments(
  TransitGatewayAttachmentIds = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{TransitGatewayAttachmentIds}{List or named list. The IDs of the attachments. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return with a single call.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes one or more VPC attachments. By default, all VPC attachments
are described. Alternatively, you can filter the results.
}
\section{TransitGatewayAttachmentIds}{

The IDs of the attachments.
}

\section{Filter}{

One or more filters. The possible values are:
\itemize{
\item \code{state} - The state of the attachment. Valid values are \code{available}
\| \code{deleted} \| \code{deleting} \| \code{failed} \| \code{failing} \|
\code{initiatingRequest} \| \code{modifying} \| \code{pendingAcceptance} \|
\code{pending} \| \code{rollingBack} \| \code{rejected} \| \code{rejecting}.
\item \code{transit-gateway-attachment-id} - The ID of the attachment.
\item \code{transit-gateway-id} - The ID of the transit gateway.
\item \code{vpc-id} - The ID of the VPC.
}
}

\section{MaxResults}{

The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

