% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_volume_attribute}
\alias{ec2_describe_volume_attribute}
\title{Describe Volume Attribute}
\usage{
ec2_describe_volume_attribute(
  Attribute,
  VolumeId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Attribute}{String. The attribute of the volume. This parameter is required.}

\item{VolumeId}{String. The ID of the volume.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the specified attribute of the specified volume. You can
specify only one attribute at a time.
}
\section{Attribute}{

The attribute of the volume. This parameter is required.
}

\section{VolumeId}{

The ID of the volume.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

