% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_vpc_classic_link}
\alias{ec2_describe_vpc_classic_link}
\title{Describe Vpc Classic Link}
\usage{
ec2_describe_vpc_classic_link(
  Filter = NULL,
  DryRun = NULL,
  VpcId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Filter}{No description can be found.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{VpcId}{Array. One or more VPCs for which you want to describe the ClassicLink status.\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes the ClassicLink status of one or
more VPCs.
}
\section{Filter}{


One or more filters.
\itemize{
\item \code{is-classic-link-enabled} - Whether the VPC is enabled for
ClassicLink (\code{true} \| \code{false}).
\item \code{tag}:\ - The key/value combination of a tag assigned to the
resource. Use the tag key in the filter name and the tag value as
the filter value. For example, to find all resources that have a tag
with the key \code{Owner} and the value \code{TeamA}, specify \code{tag:Owner} for
the filter name and \code{TeamA} for the filter value.
\item \code{tag-key} - The key of a tag assigned to the resource. Use this
filter to find all resources assigned a tag with a specific key,
regardless of the tag value.
}
}

\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{VpcId}{

One or more VPCs for which
you want to describe the ClassicLink status.
}

