% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_vpc_endpoint_connection_notifications}
\alias{ec2_describe_vpc_endpoint_connection_notifications}
\title{Describe Vpc Endpoint Connection Notifications}
\usage{
ec2_describe_vpc_endpoint_connection_notifications(
  DryRun = NULL,
  ConnectionNotificationId = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{ConnectionNotificationId}{String. The ID of the notification. (optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return in a single call.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the connection notifications for VPC endpoints and VPC
endpoint services.
}
\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{ConnectionNotificationId}{

The ID of the notification.
}

\section{Filter}{

One or more filters.
\itemize{
\item \code{connection-notification-arn} - The ARN of the SNS topic for the
notification.
\item \code{connection-notification-id} - The ID of the notification.
\item \code{connection-notification-state} - The state of the notification
(\code{Enabled} \| \code{Disabled}).
\item \code{connection-notification-type} - The type of notification (\code{Topic}).
\item \code{service-id} - The ID of the endpoint service.
\item \code{vpc-endpoint-id} - The ID of the VPC endpoint.
}
}

\section{MaxResults}{

The maximum number of results to return in a single call. To retrieve
the remaining results, make another request with the returned
\code{NextToken} value.
}

