% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_describe_vpc_endpoint_service_permissions}
\alias{ec2_describe_vpc_endpoint_service_permissions}
\title{Describe Vpc Endpoint Service Permissions}
\usage{
ec2_describe_vpc_endpoint_service_permissions(
  ServiceId,
  DryRun = NULL,
  Filter = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ServiceId}{String. The ID of the service.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{Filter}{Named list where the name is the filter name and the value is the
filter value. (see below)(optional)}

\item{MaxResults}{Integer. The maximum number of results to return for the request in a single page.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Describes the principals (service consumers) that are permitted to
discover your VPC endpoint service.
}
\section{ServiceId}{

The ID of the service.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{Filter}{

One or more filters.
\itemize{
\item \code{principal} - The ARN of the principal.
\item \code{principal-type} - The principal type (\code{All} \| \code{Service} \|
\code{OrganizationUnit} \| \code{Account} \| \code{User} \| \code{Role}).
}
}

\section{MaxResults}{

The maximum number of results to return for the request in a single
page. The remaining results of the initial request can be seen by
sending another request with the returned \code{NextToken} value. This value
can be between 5 and 1,000; if \code{MaxResults} is given a value larger than
1,000, only 1,000 results are returned.
}

