% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_detach_volume}
\alias{ec2_detach_volume}
\title{Detach Volume}
\usage{
ec2_detach_volume(
  VolumeId,
  Device = NULL,
  Force = NULL,
  InstanceId = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VolumeId}{String. The ID of the volume.}

\item{Device}{String. The device name. (optional)}

\item{Force}{Boolean. Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging...(see below)(optional)}

\item{InstanceId}{String. The ID of the instance.(optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Detaches an EBS volume from an instance. Make sure to unmount any file
systems on the device within your operating system before detaching the
volume. Failure to do so can result in the volume becoming stuck in the
\code{busy} state while detaching. If this happens, detachment can be delayed
indefinitely until you unmount the volume, force detachment, reboot the
instance, or all three. If an EBS volume is the root device of an
instance, it can\'t be detached while the instance is running. To detach
the root volume, stop the instance first.
}
\section{VolumeId}{

The ID of the volume.
}

\section{Device}{

The device name.
}

\section{Force}{

Forces detachment if the previous detachment attempt did not occur
cleanly (for example, logging into an instance, unmounting the volume,
and detaching normally). This option can lead to data loss or a
corrupted file system. Use this option only as a last resort to detach a
volume from a failed instance. The instance won\'t have an opportunity
to flush file system caches or file system metadata. If you use this
option, you must perform file system check and repair procedures.
}

\section{InstanceId}{

The ID of the instance. If you are detaching a Multi-Attach enabled
volume, you must specify an instance ID.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

