% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disable_vpc_classic_link_dns_support}
\alias{ec2_disable_vpc_classic_link_dns_support}
\title{Disable Vpc Classic Link Dns Support}
\usage{
ec2_disable_vpc_classic_link_dns_support(
  VpcId = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcId}{String. The ID of the VPC.\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Disable Vpc Classic Link Dns Support
}
\section{VpcId}{

The ID of the
VPC.
}

