% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disassociate_address}
\alias{ec2_disassociate_address}
\title{Disassociate Address}
\usage{
ec2_disassociate_address(
  AssociationId = NULL,
  PublicIp = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. [EC2-VPC] The association ID. Required for EC2-VPC. (optional)}

\item{PublicIp}{String. [EC2-Classic] The Elastic IP address. Required for EC2-Classic. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Disassociates an Elastic IP address from the instance or network
interface it\'s associated with.
}
\section{AssociationId}{

[EC2-VPC] The association ID. Required for EC2-VPC.
}

\section{PublicIp}{

[EC2-Classic] The Elastic IP address. Required for EC2-Classic.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

