% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disassociate_subnet_cidr_block}
\alias{ec2_disassociate_subnet_cidr_block}
\title{Disassociate Subnet Cidr Block}
\usage{
ec2_disassociate_subnet_cidr_block(
  AssociationId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. The association ID for the CIDR block.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Disassociates a CIDR block from a subnet.
Currently, you can disassociate an IPv6 CIDR block only. You must detach
or delete all gateways and resources that are associated with the CIDR
block before you can disassociate it.
}
\section{AssociationId}{

The association ID for the
CIDR block.
}

