% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_disassociate_vpc_cidr_block}
\alias{ec2_disassociate_vpc_cidr_block}
\title{Disassociate Vpc Cidr Block}
\usage{
ec2_disassociate_vpc_cidr_block(
  AssociationId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. The association ID for the CIDR block.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Disassociates a CIDR block from a VPC. To disassociate the CIDR block,
you must specify its association ID. You can get the association ID by
using DescribeVpcs. You must detach or delete all gateways and resources
that are associated with the CIDR block before you can disassociate it.
}
\section{AssociationId}{

The association ID for the CIDR block.
}

