% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_export_client_vpn_client_certificate_revocation_list}
\alias{ec2_export_client_vpn_client_certificate_revocation_list}
\title{Export Client Vpn Client Certificate Revocation List}
\usage{
ec2_export_client_vpn_client_certificate_revocation_list(
  ClientVpnEndpointId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ClientVpnEndpointId}{String. The ID of the Client VPN endpoint.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Downloads the client certificate revocation list for the specified
Client VPN endpoint.
}
\section{ClientVpnEndpointId}{

The ID of the Client VPN endpoint.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

