% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_console_output}
\alias{ec2_get_console_output}
\title{Get Console Output}
\usage{
ec2_get_console_output(
  InstanceId,
  DryRun = NULL,
  Latest = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{String. The ID of the instance.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{Latest}{Logical.   When enabled, retrieves the latest console output for the instance.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Get Console Output
}
\section{InstanceId}{

The ID of the
instance.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{Latest}{


When enabled, retrieves the latest console output for the instance.

Default: disabled (\code{false})
}

