% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_get_host_reservation_purchase_preview}
\alias{ec2_get_host_reservation_purchase_preview}
\title{Get Host Reservation Purchase Preview}
\usage{
ec2_get_host_reservation_purchase_preview(
  HostIdSet,
  OfferingId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{HostIdSet}{List or named list. The IDs of the Dedicated Hosts with which the reservation is associated.}

\item{OfferingId}{String. The offering ID of the reservation.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Preview a reservation purchase with configurations that match those of
your Dedicated Host. You must have active Dedicated Hosts in your
account before you purchase a reservation.
}
\section{HostIdSet}{

The IDs of the Dedicated Hosts with which the reservation is associated.
}

\section{OfferingId}{

The offering ID of the reservation.
}

