% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_import_key_pair}
\alias{ec2_import_key_pair}
\title{Import Key Pair}
\usage{
ec2_import_key_pair(
  KeyName,
  PublicKeyMaterial,
  DryRun = NULL,
  TagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{KeyName}{String. A unique name for the key pair.}

\item{PublicKeyMaterial}{String. The public key.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{TagSpecification}{List or named list. The tags to apply to the imported key pair. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Imports the public key from an RSA key pair that you created with a
third-party tool. Compare this with CreateKeyPair, in which AWS creates
the key pair and gives the keys to you (AWS keeps a copy of the public
key). With ImportKeyPair, you create the key pair and give AWS just the
public key. The private key is never transferred between you and AWS.
}
\section{KeyName}{

A unique name for the key pair.
}

\section{PublicKeyMaterial}{

The public key. For API calls, the text must be base64-encoded. For
command line tools, base64 encoding is performed for you.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{TagSpecification}{

The tags to apply to the imported key pair.
}

