% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_import_volume}
\alias{ec2_import_volume}
\title{Import Volume}
\usage{
ec2_import_volume(
  AvailabilityZone,
  Image,
  Volume,
  Description = NULL,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AvailabilityZone}{String. The Availability Zone for the resulting EBS volume.}

\item{Image}{Object. The disk image.}

\item{Volume}{Object. The volume size.}

\item{Description}{String. A description of the volume. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Creates an import volume task using metadata from the specified disk
image.For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/CommandLineReference/importing-your-volumes-into-amazon-ebs.html}{Importing Disks to Amazon EBS}.
}
\section{AvailabilityZone}{

The Availability Zone for the resulting EBS volume.
}

\section{Image}{

The disk image.
}

\section{Volume}{

The volume size.
}

\section{Description}{

A description of the volume.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

