% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_fleet}
\alias{ec2_modify_fleet}
\title{Modify Fleet}
\usage{
ec2_modify_fleet(
  FleetId,
  DryRun = NULL,
  ExcessCapacityTerminationPolicy = NULL,
  LaunchTemplateConfig = NULL,
  TargetCapacitySpecification = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{FleetId}{Character. The ID of the EC2 Fleet.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{ExcessCapacityTerminationPolicy}{Character. Indicates whether running instances should be terminated if the total target capacity of the EC2...[optional]}

\item{LaunchTemplateConfig}{List. The launch template and overrides.[optional]}

\item{TargetCapacitySpecification}{Object. The size of the EC2 Fleet.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Modify Fleet
}
\section{FleetId}{

The ID of the EC2
Fleet.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{ExcessCapacityTerminationPolicy}{

Indicates whether running
instances should be terminated if the total target capacity of the EC2
Fleet is decreased below the current size of the EC2
Fleet.
}

\section{LaunchTemplateConfig}{

The launch template and
overrides.
}

\section{TargetCapacitySpecification}{

The size of the EC2
Fleet.
}

