% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_id_format}
\alias{ec2_modify_id_format}
\title{Modify Id Format}
\usage{
ec2_modify_id_format(Resource, UseLongIds, simplify = TRUE, others = list())
}
\arguments{
\item{Resource}{String.   The type of resource: \code{bundle} \| \code{conversion-task} \| \code{customer-gateway} \| \code{dhcp-options}...}

\item{UseLongIds}{Logical. Indicate whether the resource should use longer IDs (17-character IDs).}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Modify Id Format
}
\section{Resource}{


The type of resource: \code{bundle} \| \code{conversion-task} \|
\code{customer-gateway} \| \code{dhcp-options} \| \code{elastic-ip-allocation} \|
\code{elastic-ip-association} \| \code{export-task} \| \code{flow-log} \| \code{image} \|
\code{import-task} \| \code{internet-gateway} \| \code{network-acl} \|
\code{network-acl-association} \| \code{network-interface} \|
\code{network-interface-attachment} \| \code{prefix-list} \| \code{route-table} \|
\code{route-table-association} \| \code{security-group} \| \code{subnet} \|
\code{subnet-cidr-block-association} \| \code{vpc} \| \code{vpc-cidr-block-association}
\| \code{vpc-endpoint} \| \code{vpc-peering-connection} \| \code{vpn-connection} \|
\code{vpn-gateway}.

Alternatively, use the \code{all-current} option to include all resource
types that are currently within their opt-in period for longer IDs.
}

\section{UseLongIds}{

Indicate whether the resource should use longer IDs
(17-character IDs).
}

