% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_instance_event_start_time}
\alias{ec2_modify_instance_event_start_time}
\title{Modify Instance Event Start Time}
\usage{
ec2_modify_instance_event_start_time(
  InstanceId,
  InstanceEventId,
  NotBefore,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{InstanceId}{String. The ID of the instance with the scheduled event.}

\item{InstanceEventId}{String. The ID of the event whose date and time you are modifying.}

\item{NotBefore}{String. The new date and time when the event will take place.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...\link{optional}}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Modifies the start time for a scheduled Amazon EC2
instance event.
}
\section{InstanceId}{

The
ID of the instance with the scheduled event.
}

\section{InstanceEventId}{

The ID
of the event whose date and time you are
modifying.
}

\section{NotBefore}{

The new date and time when the event will
take place.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

