% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_managed_prefix_list}
\alias{ec2_modify_managed_prefix_list}
\title{Modify Managed Prefix List}
\usage{
ec2_modify_managed_prefix_list(
  PrefixListId,
  DryRun = NULL,
  CurrentVersion = NULL,
  PrefixListName = NULL,
  AddEntry = NULL,
  RemoveEntry = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout(),
  region = aws_get_region()
)
}
\arguments{
\item{PrefixListId}{Character. The ID of the prefix list.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{CurrentVersion}{Integer. The current version of the prefix list.[optional]}

\item{PrefixListName}{Character. A name for the prefix list.[optional]}

\item{AddEntry}{List. One or more entries to add to the prefix list.[optional]}

\item{RemoveEntry}{List. One or more entries to remove from the prefix list.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}

\item{region}{Character. The region of the AWS service.}
}
\value{
A list object or a character vector
}
\description{
Modify Managed Prefix List
}
\section{PrefixListId}{

The ID of the prefix
list.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{CurrentVersion}{

The current version of the
prefix list.
}

\section{PrefixListName}{

A name for the prefix
list.
}

\section{AddEntry}{

One or more entries to add to the prefix
list.
}

\section{RemoveEntry}{

One or more entries to remove from the prefix
list.
}

