% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_reserved_instances}
\alias{ec2_modify_reserved_instances}
\title{Modify Reserved Instances}
\usage{
ec2_modify_reserved_instances(
  ReservedInstancesId,
  ReservedInstancesConfigurationSetItemType,
  ClientToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ReservedInstancesId}{List or named list. The IDs of the Reserved Instances to modify.}

\item{ReservedInstancesConfigurationSetItemType}{List or named list. The configuration settings for the Reserved Instances to modify.}

\item{ClientToken}{String. A unique, case-sensitive token you provide to ensure idempotency of your modification request.(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the Availability Zone, instance count, instance type, or
network platform (EC2-Classic or EC2-VPC) of your Reserved Instances.
The Reserved Instances to be modified must be identical, except for
Availability Zone, network platform, and instance type.
}
\section{ReservedInstancesId}{

The IDs of the Reserved Instances to modify.
}

\section{ReservedInstancesConfigurationSetItemType}{

The configuration settings for the Reserved Instances to modify.
}

\section{ClientToken}{

A unique, case-sensitive token you provide to ensure idempotency of your
modification request. For more information, see \href{https://docs.aws.amazon.com/AWSEC2/latest/APIReference/Run_Instance_Idempotency.html}{Ensuring Idempotency}.
}

