% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_vpc_endpoint_service_permissions}
\alias{ec2_modify_vpc_endpoint_service_permissions}
\title{Modify Vpc Endpoint Service Permissions}
\usage{
ec2_modify_vpc_endpoint_service_permissions(
  ServiceId,
  DryRun = NULL,
  AddAllowedPrincipals = NULL,
  RemoveAllowedPrincipals = NULL,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout(),
  region = aws_get_region()
)
}
\arguments{
\item{ServiceId}{Character. The ID of the service.}

\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the request,...[optional]}

\item{AddAllowedPrincipals}{List. The Amazon Resource Names (ARN) of one or more principals.[optional]}

\item{RemoveAllowedPrincipals}{List. The Amazon Resource Names (ARN) of one or more principals.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}

\item{region}{Character. The region of the AWS service.}
}
\value{
A list object or a character vector
}
\description{
Modify Vpc Endpoint Service Permissions
}
\section{ServiceId}{

The ID of the
service.
}

\section{DryRun}{

Checks whether you have the required
permissions for the action, without actually making the request, and
provides an error response. If you have the required permissions, the
error response is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{AddAllowedPrincipals}{

The Amazon Resource Names
(ARN) of one or more principals. Permissions are granted to the
principals in this list. To grant permissions to all principals, specify
an asterisk (\*).
}

\section{RemoveAllowedPrincipals}{

The Amazon Resource Names (ARN) of
one or more principals. Permissions are revoked for principals in this
list.
}

