% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_modify_vpc_tenancy}
\alias{ec2_modify_vpc_tenancy}
\title{Modify Vpc Tenancy}
\usage{
ec2_modify_vpc_tenancy(
  VpcId,
  InstanceTenancy,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{VpcId}{String. The ID of the VPC.}

\item{InstanceTenancy}{String. The instance tenancy attribute for the VPC.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Modifies the instance tenancy attribute of the specified VPC. You can
change the instance tenancy attribute of a VPC to \code{default} only. You
cannot change the instance tenancy attribute to \code{dedicated}.
}
\section{VpcId}{

The ID of the VPC.
}

\section{InstanceTenancy}{

The instance tenancy attribute for the VPC.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

