% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_provision_byoip_cidr}
\alias{ec2_provision_byoip_cidr}
\title{Provision Byoip Cidr}
\usage{
ec2_provision_byoip_cidr(
  Cidr,
  CidrAuthorizationContext = NULL,
  PubliclyAdvertisable = NULL,
  Description = NULL,
  DryRun = NULL,
  PoolTagSpecification = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Cidr}{String. The public IPv4 or IPv6 address range, in CIDR notation.}

\item{CidrAuthorizationContext}{Object. A signed document that proves that you are authorized to bring the specified IP address range...(see below)(optional)}

\item{PubliclyAdvertisable}{Boolean. (IPv6 only) Indicate whether the address range will be publicly advertised to the internet.(optional)}

\item{Description}{String. A description for the address range and the address pool. (optional)}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{PoolTagSpecification}{List or named list. The tags to apply to the address pool. (optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Provisions an IPv4 or IPv6 address range for use with your AWS resources
through bring your own IP addresses (BYOIP) and creates a corresponding
address pool. After the address range is provisioned, it is ready to be
advertised using AdvertiseByoipCidr.
}
\section{Cidr}{

The public IPv4 or IPv6 address range, in CIDR notation. The most
specific IPv4 prefix that you can specify is /24. The most specific IPv6
prefix you can specify is /56. The address range cannot overlap with
another address range that you\'ve brought to this or another Region.
}

\section{CidrAuthorizationContext}{

A signed document that proves that you are authorized to bring the
specified IP address range to Amazon using BYOIP.
}

\section{PubliclyAdvertisable}{

(IPv6 only) Indicate whether the address range will be publicly
advertised to the internet.

Default: true
}

\section{Description}{

A description for the address range and the address pool.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

\section{PoolTagSpecification}{

The tags to apply to the address pool.
}

