% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_register_instance_event_notification_attributes}
\alias{ec2_register_instance_event_notification_attributes}
\title{Register Instance Event Notification Attributes}
\usage{
ec2_register_instance_event_notification_attributes(
  DryRun = NULL,
  InstanceTagAttribute = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{DryRun}{Logical. Checks whether you have the required permissions for the action, without actually making the...[optional]}

\item{InstanceTagAttribute}{Object. Information about the tag keys to register.[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Register Instance Event Notification Attributes
}
\section{DryRun}{

Checks whether you have the required permissions
for the action, without actually making the request, and provides an
error response. If you have the required permissions, the error response
is \code{DryRunOperation}. Otherwise, it is
\code{UnauthorizedOperation}.
}

\section{InstanceTagAttribute}{

Information about the tag
keys to register.
}

