% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_release_hosts}
\alias{ec2_release_hosts}
\title{Release Hosts}
\usage{
ec2_release_hosts(HostId, simplify = TRUE, others = list())
}
\arguments{
\item{HostId}{Array. The IDs of the Dedicated Hosts to release.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Release Hosts
}
\section{HostId}{

The IDs of the Dedicated
Hosts to release.
}

