% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_replace_iam_instance_profile_association}
\alias{ec2_replace_iam_instance_profile_association}
\title{Replace Iam Instance Profile Association}
\usage{
ec2_replace_iam_instance_profile_association(
  IamInstanceProfile,
  AssociationId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{IamInstanceProfile}{Object. The IAM instance profile.}

\item{AssociationId}{Character. The ID of the existing IAM instance profile association.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Replace Iam Instance Profile Association
}
\section{IamInstanceProfile}{

The IAM instance
profile.
}

\section{AssociationId}{

The ID of the existing IAM instance profile
association.
}

