% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_replace_iam_instance_profile_association}
\alias{ec2_replace_iam_instance_profile_association}
\title{Replace Iam Instance Profile Association}
\usage{
ec2_replace_iam_instance_profile_association(
  IamInstanceProfile,
  AssociationId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{IamInstanceProfile}{Object. The IAM instance profile.}

\item{AssociationId}{String. The ID of the existing IAM instance profile association.}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Replaces an IAM instance profile for the specified running instance. You
can use this action to change the IAM instance profile that\'s
associated with an instance without having to disassociate the existing
IAM instance profile first.
}
\section{IamInstanceProfile}{

The IAM instance profile.
}

\section{AssociationId}{

The ID of the existing IAM instance profile association.
}

