% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_replace_route_table_association}
\alias{ec2_replace_route_table_association}
\title{Replace Route Table Association}
\usage{
ec2_replace_route_table_association(
  AssociationId,
  RouteTableId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{AssociationId}{String. The association ID.}

\item{RouteTableId}{String. The ID of the new route table to associate with the subnet.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Changes the route table associated with a given subnet, internet
gateway, or virtual private gateway in a VPC. After the operation
completes, the subnet or gateway uses the routes in the new route table.
For more information about route tables, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/VPC_Route_Tables.html}{Route Tables}
in the \emph{Amazon Virtual Private Cloud User Guide}.
}
\section{AssociationId}{

The association ID.
}

\section{RouteTableId}{

The ID of the new route table to associate with the subnet.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

