% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_start_vpc_endpoint_service_private_dns_verification}
\alias{ec2_start_vpc_endpoint_service_private_dns_verification}
\title{Start Vpc Endpoint Service Private Dns Verification}
\usage{
ec2_start_vpc_endpoint_service_private_dns_verification(
  ServiceId,
  DryRun = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{ServiceId}{String. The ID of the endpoint service.}

\item{DryRun}{Boolean. Checks whether you have the required permissions for the action, without actually making the...(see below)(optional)}

\item{simplify}{Whether to handle the nextToken value(if any)
in the response and simplify the result.}
}
\value{
A list object or a character vector
}
\description{
Initiates the verification process to prove that the service provider
owns the private DNS name domain for the endpoint service.
}
\section{ServiceId}{

The ID of the endpoint service.
}

\section{DryRun}{

Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.
}

