% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_unassign_ipv6_addresses}
\alias{ec2_unassign_ipv6_addresses}
\title{Unassign Ipv6 Addresses}
\usage{
ec2_unassign_ipv6_addresses(
  Ipv6Addresses,
  NetworkInterfaceId,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout(),
  region = aws_get_region()
)
}
\arguments{
\item{Ipv6Addresses}{List. The IPv6 addresses to unassign from the network interface.}

\item{NetworkInterfaceId}{Character. The ID of the network interface.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}

\item{region}{Character. The region of the AWS service.}
}
\value{
A list object or a character vector
}
\description{
Unassigns one or more IPv6 addresses from a network
interface.
}
\section{Ipv6Addresses}{

The IPv6 addresses to
unassign from the network interface.
}

\section{NetworkInterfaceId}{

The ID of the
network interface.
}

