% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_unassign_ipv6_addresses}
\alias{ec2_unassign_ipv6_addresses}
\title{Unassign Ipv6 Addresses}
\usage{
ec2_unassign_ipv6_addresses(
  Ipv6Addresses,
  NetworkInterfaceId,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{Ipv6Addresses}{List. The IPv6 addresses to unassign from the network interface.}

\item{NetworkInterfaceId}{String. The ID of the network interface.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Unassigns one or more IPv6 addresses from a network
interface.
}
\section{Ipv6Addresses}{

The IPv6 addresses to
unassign from the network interface.
}

\section{NetworkInterfaceId}{

The ID of the
network interface.
}

