% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-actions.R
\name{ec2_unassign_private_ip_addresses}
\alias{ec2_unassign_private_ip_addresses}
\title{Unassign Private Ip Addresses}
\usage{
ec2_unassign_private_ip_addresses(
  NetworkInterfaceId,
  PrivateIpAddress,
  simplify = TRUE,
  others = list(),
  print_on_error = aws_get_print_on_error(),
  retry_time = aws_get_retry_time(),
  network_timeout = aws_get_network_timeout()
)
}
\arguments{
\item{NetworkInterfaceId}{Character. The ID of the network interface.}

\item{PrivateIpAddress}{List. The secondary private IP addresses to unassign from the network interface.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}

\item{print_on_error}{Logical. Whether to show an error message when a
network error occurs.}

\item{retry_time}{Integer. Number of retries for a REST request when encounter the
network issue. If the request has been sent \code{retry_time} times but still not
be able to get the response, an error will be thrown.}

\item{network_timeout}{Numeric. Number of seconds to wait for a REST response
until giving up. Can not be less than 1 ms.}
}
\value{
A list object or a character vector
}
\description{
Unassigns one or more secondary private
IP addresses from a network interface.
}
\section{NetworkInterfaceId}{

The ID of the network
interface.
}

\section{PrivateIpAddress}{

The secondary private IP addresses to
unassign from the network interface. You can specify this option
multiple times to unassign more than one IP
address.
}

