% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_deregister_task_definition}
\alias{ecs_deregister_task_definition}
\title{Deregister Task Definition}
\usage{
ecs_deregister_task_definition(
  taskDefinition = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{taskDefinition}{String. The \code{family} and \code{revision} (\code{family:revision}) or full Amazon Resource Name (ARN) of the...}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Deregister Task Definition
}
\section{taskDefinition}{

The \code{family} and \code{revision} (\code{family:revision}) or
full Amazon Resource Name (ARN) of the task definition to deregister.
You must specify a \code{revision}.
}

