% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_describe_container_instances}
\alias{ecs_describe_container_instances}
\title{Describe Container Instances}
\usage{
ecs_describe_container_instances(
  cluster = NULL,
  containerInstances = NULL,
  include = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{cluster}{String. The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container...}

\item{containerInstances}{List. A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.}

\item{include}{List. Specifies whether you want to see the resource tags for the container instance.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Describes Amazon Elastic Container Service
container instances. Returns metadata about registered and remaining
resources on each container instance
requested.
}
\section{cluster}{

The short name or full Amazon Resource Name (ARN)
of the cluster that hosts the container instances to describe. If you do
not specify a cluster, the default cluster is assumed. This parameter is
required if the container instance or container instances you are
describing were launched in any cluster other than the default
cluster.
}

\section{containerInstances}{

A list of up to 100 container instance IDs
or full Amazon Resource Name (ARN)
entries.
}

\section{include}{

Specifies whether you want to see the
resource tags for the container instance. If \code{TAGS} is specified, the
tags are included in the response. If this field is omitted, tags are
not included in the response.
}

