% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_describe_container_instances}
\alias{ecs_describe_container_instances}
\title{Describe Container Instances}
\usage{
ecs_describe_container_instances(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Describes Amazon Elastic Container Service
container instances. Returns metadata about registered and remaining
resources on each container instance
requested.
}
