% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_list_services}
\alias{ecs_list_services}
\title{List Services}
\usage{
ecs_list_services(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Lists the services that are running in a
specified cluster.
}
\section{maxResults}{

Pagination
limit
}

