% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_list_task_definitions}
\alias{ecs_list_task_definitions}
\title{List Task Definitions}
\usage{
ecs_list_task_definitions(
  maxResults = NULL,
  nextToken = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{maxResults}{String. Pagination limit[optional]}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Returns a list of task definitions that are
registered to your account. You can filter the results by family name
with the \code{familyPrefix} parameter or by status with the \code{status}
parameter.
}
\section{maxResults}{

Pagination
limit
}

