% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_untag_resource}
\alias{ecs_untag_resource}
\title{Untag Resource}
\usage{
ecs_untag_resource(
  resourceArn = NULL,
  tagKeys = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{resourceArn}{Character. The Amazon Resource Name (ARN) of the resource from which to delete tags.}

\item{tagKeys}{List. The keys of the tags to be removed.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Deletes specified tags from a
resource.
}
\section{resourceArn}{

The Amazon Resource Name (ARN) of the resource from
which to delete tags. Currently, the supported resources are Amazon ECS
capacity providers, tasks, services, task definitions, clusters, and
container instances.
}

\section{tagKeys}{

The keys of the tags to be
removed.
}

