% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_update_capacity_provider}
\alias{ecs_update_capacity_provider}
\title{Update Capacity Provider}
\usage{
ecs_update_capacity_provider(
  name = NULL,
  autoScalingGroupProvider = NULL,
  simplify = TRUE,
  others = list()
)
}
\arguments{
\item{name}{Character. The name of the capacity provider to update.}

\item{autoScalingGroupProvider}{Object. An object representing the parameters to update for the Auto Scaling group capacity provider.}

\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response[optional]}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request[optional]}
}
\value{
A list object or a character vector
}
\description{
Modifies the parameters for a capacity
provider.
}
\section{name}{

The name of the capacity provider to
update.
}

\section{autoScalingGroupProvider}{

An object representing the parameters to
update for the Auto Scaling group capacity
provider.
}

