% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs-actions.R
\name{ecs_update_service_primary_task_set}
\alias{ecs_update_service_primary_task_set}
\title{Update Service Primary Task Set}
\usage{
ecs_update_service_primary_task_set(simplify = TRUE, others = list())
}
\arguments{
\item{simplify}{Logical. Whether to simplify the result and handle \code{nextToken}
in the response}

\item{others}{Named list. The parameters that are not included in the
function parameters and need to be added into the request.}
}
\value{
A list object or a character vector
}
\description{
Modifies which task set in a service is the primary
task set. Any parameters that are updated on the primary task set in a
service will transition to the service. This is used when a service uses
the \code{EXTERNAL} deployment controller type. For more information, see
\href{https://docs.aws.amazon.com/AmazonECS/latest/developerguide/deployment-types.html}{Amazon ECS Deployment Types}
in the \emph{Amazon Elastic Container Service Developer
Guide}.
}
